run tests in isolated environments
http://github.com/astriaorg/sequencer-relayer/pull/24

This commit introduces isolated environments for
all tests that require external services.

This is done by using podman to spin up a fresh
pod consiting of the celestia-app, celestia-node
(as a bridge to celestia), and metro containers.
In order for the exposed pod ports to not clash,
a new kube yaml is generated from the template at
`templates/sequencer_relayer_stack.yaml.jinja2`
and sent to the podman API service (see the
README for more information about how to install
podman and start the API service).
In addition, pods and emptyDir mounts receive
unique per-test names to work around podman
idiosyncracies.

The pods are stopped and removed when the tests
go out of scope.

On the CI side, podman version 4+ needs to be
installed from the suse kubic repos, which at
the time of writing are still maintained. podman
version 3 shipped with ubuntu-22.04 used
by the github runners is too old to support
all podman-kube-play options.