feat: implement celestia json rpc client
This commit replaces `astria-rs-cnc` with the
`astria-celestia-jsonrpc-client` crate, and changes
astria-sequencer-relayer and astria-conductor to
work in terms of it.

The only endpoints currenly implemented are those
used by astria-sequencer-relayer and astria-conductor:

+ blob.Submit (to submit one or more blobs)
+ blob.GetAll (to get all blobs given a height and a namespace ID)
+ state.SubmitPayForBlob (to submit one or more blobs with payment)
+ header.NetworkHead (to query the latest network head and hence height)

Because all JSON RPCs are protected with bearer tokens,
these are now generated after celestia-node is brought up
in the test pods using a lifecycle hook. The token is then
extracted using the kubernetes REST API in each test.

As part of this commit the test environments for conductor
and sequencer-relayer were bumped to celestia-app:v1.0.0-rc7
and celestia-node:v0.11.0-rc7.

fix readme

address comments