container: install dynamic libs and fix entrypoint
This commit installs rocksdb in the final layer
of the generic containerfile because all binaries
are currently linked to librocksdb.

In addition, as it turns out that build args are not
expanded in ENTRYPOINT lines. To get a stable path
that can be used as an argument to ENTRYPOINT, this
commit crates a softlink to /usr/bin/local/$TARGETBINARY
and uses that instead.