sequencer-relayer: more sophisticated telemetry setup (#163)
This commit changes telemetry to output a human readable format only if connected to a tty, and json otherwise. It also allows configuring env filters via command line arguments in addition to environment variables.

This commit was broken out of #150. It was time to upgrade the tracing setup either way, so it made sense to make it a separate PR.