proto: commit generated code to repository (#188)
This commit checks in the rust code compiled from
the protobuf definitions. This helps with:

1. better discovery of the compiled rust types by simply
navigating to the files;
2. better usability for users who just want to use the rust
crates because they now need not install buf to compile
the protobufs to rust on every build.

While it is convention to not commit generated
code to the repository, I believe the two points
above are good arguments why one should do this
in this case. Also, the protobuf definitions are
not expected to change frequently.

In addition, this commit pins the version of the
remote buf plugins to defend against accidental
breakage.