update conductor to use telemetry
To make conductor work with the standardized
telemetry setup, its `Config` had to be extended
by a `log` field. To not cause breaking changes
with its previous telemetry setup, a figment
env provider was added that only reads
`RUST_LOG`. If `Config.log` is not set through
any of the providers, it falls back to
a global `info` directive, as before.