ci: ensure that crates compile outside of workspace (#259)
## Summary (why this PR?)

#239 added the secrecy crate to composer, which builds fine if cargo is
executed from the workspace root, but fails to build if run from the
crate manifest directory. This is because optional dependencies get
automatically activated if the dependency is used in any other crate or
dependency. This is due to how rust dependency resolution works.

## Changelist

+ add a CI workflow step running `cargo hack check --all` which ensures
that all crates build without workspace-wide dependency resolution.

## Breaking changes

None

## Related issues

+ https://github.com/astriaorg/astria/pull/239 (broke compilation with working directory set to `crates/astria-composer/`)

## Testing

None (CI either works or does not)