composer: ingest transactions from multiple rollups (#278)
This is second PR of three PRs to support reading from multiple rollups
in composer.

This patch defines the `Collector` type as a long running tasks that
read transactions from a different rollup each and submit them to the
searcher together with a chain ID. Because the searcher has to keep
track of the nonce, it still converts and signs the different
transactions when converting them to sequencer blocks.

This patch enables https://github.com/astriaorg/astria/issues/214

- Define the `Collector` type, which reads new pending transactions from
an EVM rollup supporting the ethereum execution API.
- Replace the the new pending transaction stream in searcher by an mpsc
channel that receives from a set of collector.
- Spin up one collector per `<chain_id>::<url>` pair provided through
config.
- Each collector tags its transaction with its unique chain ID.

The blackbox tests were extended to read from 2 different collectors
implemented in terms of two separate mock servers.

Sibling PRs:

- https://github.com/astriaorg/astria/pull/277
- https://github.com/astriaorg/astria/pull/278 (the present PR)
- https://github.com/astriaorg/astria/pull/282

- https://github.com/astriaorg/astria/issues/214 (will be closed after
issue 277 is merged)