introduce ChainId native type an implement everything in terms of it
The introduction of the `ChainId` type had ripple effects on several
parts of the codebase:

1. The proof construction could be expressed in terms of [u8; 32] instead of
Vec<u8>, saving a lot of allocations;
2. This made it necessary to make the `MerkleTree` abstract over `T: AsRef<[u8]>`, because
it's now used with both types internally.
3. Unfortunately the abstraction is leaky because `CtMerkle` has a trait bound on its
type definition, leaking through to our definition of `MerkleTree`
4. The `Default` impl for `RollupData` was removed because it has no natural default.
Namespace is constructed as a sha256 over bytes, and a 0-hash is a valid hash.
5. this triggered changes in conductor: the SequencerBlockSubset constructor now returns
an Option.
6. if None, the subset is skipped and not executed.
7. this does not change functionality because during execution subsets are checked for empty data
and skipped