ci: ensure Cargo.lock is up to date (#315)
## Summary

Add a job to the tests workflow to ensure the cargo lockfile is updated.

## Background

If a change is made to the `Cargo.toml` of a workspace member without
running `cargo {fetch,check,build,run}` (among others), then the
workspace `Cargo.lock` file is not updated. This means the lockfile is
outdated and does not match the lockfile that will be used when building
artifacts. This can have negative effects on caching and build times in
CI.

## Changes

- Added a new `lockfile` job to the `test.yml` workflow.
- Made `lockfile` job a requirement for workflow success.