proto: define sequencer v1alpha2
This patch extracts Rust types that are used in communication between
sequencer, sequencer-relayer, conductor, and celestia (as the
data availability layer). The resulting protobuf types are intended to
be the only way that services exchange data, making the current json-based
wire format obsolete.

The protobuf types defined in v1alpha1 of the sequencer API are deprecated
(and were also never used).