remove gossipnet from sequencer-relayer
This also removes the flag to disable writing to the DA layer by sequencer-relayer.
This is now its only functionality, so it will always attempt to write to celestia.

Removing gossipnet changed how the individual steps of sequener-relayer
had to be controlled to check if blocks were indeed submitted to
celestia as a batch. It also revealed some underlying issues with
sequencer requests running indefinitely. This change times out
requests to get new blocks from sequencer after 2 times the poll
period have passed.