gossipnet: remove (#407)
## Summary
Remove the gossipnet crate and mentions of it.

## Background
Patches https://github.com/astriaorg/astria/pull/383,
https://github.com/astriaorg/astria/pull/384 removed gossipnet from
sequencer-relayer and conductor. This patch removes the
`astria-gossipnet` crate altogether because it is no longer used.

This PR does not update the conductor spec but leaves it as a follow-up.

## Changes
- Remove `astria-gossipnet`, update README, github workflow, workspace
Cargo.toml to reflect that
- Remove `gossipnet` mentions from astria-conductor

## Testing
not applicable

## Related Issues

closes https://github.com/astriaorg/astria/issues/334