feat(celestia-client): use eigerco celestia node client
This patch removes astria-celestia-jsonrpc-client in favor
of the client derived in github.com/eigerco/celestia-node-rs.

The `data_availability` module defined in sequencer-relayer
is also removed in favor of an extension trait in the new
`astria-celestia-client` that makes it more ergonomic to
submit sequencer blocks to celestia and reade them back.

This means there are no more direct dependencies of
astria services on any other.

This patch also adds a new crate `astria-celestia-mock`
that for now only contains RPC server traits for creating
a celestia-node mock in tests.

All tests requiring kind/kubernetes are removed from CI.

Protoc was added to CI and to the containerfile because
of new dependencies that require it.