conductor: reconnect to sequencer websocket with backoff (#483)
## Summary
After losing its websocket connection to sequener conductor now tries to
reconnect up to 1024 times with exponential backoff.

## Background
Conductor only tried to reestablish its websocket connection to
sequencer up to one time, bailing afterwards. This is not enough to
overcome network issues, leading to conductor being unnecessarily flaky
and bailing too early.

## Changes
- use exponential backoff to reestablish the sequencer websocket
connection (up to 1024 attempts)
- merge the startup and reconnect logic as they are essentially the same
thing.

## Testing
Testing this is tough and not done here. This is tough to implement and
was not done here. Issue to track this:
https://github.com/astriaorg/astria/issues/482