conductor: spawn driver as task and report exit
Spawns the tendermint websocket driver as a
dedicated task and reports why it exited/failed.

The websocket driver should immediately start
communicating with the remote host, sending keep-alives
etc.

Also, the tendermint websocket driver frequently disconnects, but we are
ignoring its return value. We should not do that because it might
contain valuable information about why it exited.

- Spawn the websocket driver as a task
- Add error message and backtrace to event emitted after websocket
driver exited, but before it returned.