fix(sequencer): move transaction execution to prepare/process proposal (#480)
## Summary
We don't want transactions (specifically `sequence::Action`s) which fail
fee payment to end up in the block data, as they would essentially get
sequenced for free. The easiest way to deal with this is to execute all
transactions in a block *before* the block is committed by the
tendermint network, and only include txs that successfully execute,
which is what this PR does.

## Background
see #478

## Changes
- update `prepare_proposal` and `process_proposal` to execute
transactions instead of later on in the usual `deliver_tx` loop.
- the results of tx execution are cached in the `App` and returned when
`deliver_tx` is called later.
- as a result, `begin_block` is actually executed *after* all the
transactions are executed. this is fine, since the only state change
`begin_block` can do at the moment is removal of byzantine validators
(plus updating block height + timestamp), none of which affect tx
execution.

## Testing
unit tests

## Breaking Changelist
any tx which fails execution will no longer be included in a block's tx
data.

## Related Issues

closes #478