feat(conductor): add re-sync for missed sequencer blocks (#515)
## Summary
Fetch missing sequencer blocks before forwarding to the executor actor.

## Background
If the websocket connection to sequencer drops temporarily (thus not
receiving blocks from the `NewBlocks` jsonrpc subscription), blocks for
one or more heights might be missed. This is detected by tracking the
expected height of the next sequencer block to be forwarded to the
executor actor. If the next block taken from the queue of pending blocks
has a height greater than the expected height a resync is triggered and
the current block rescheduled in front of the queue of regular pending
blocks. This patch reuses the sync-mechanism that is run at startup to
fetch a contiguous range of missing sequencer blocks. Only after this
"re-sync" is completed will regular submission of blocks resume

## Changes
- Track the height of sequencer blocks
- Trigger a resync if there is a mismatch between the height in the next
sequencer block and the expected height
- Resume submission of blocks after resync is done

## Testing
Added unit tests for all three forward/resync conditions:
1. forward to executor if block height == expected height
2. drop if block height < expected height
3. sync + reschedule if block height > expected height

No blackbox tests yet. Must be tested on devnet.