composer: resubscribe if rollup subscription stops (#532)
## Summary
Added logic to reconnect and resubscribe to a rollup if the subscription
ends.

## Background
Composer exited if a websocket/jsonrpc subscription to a remote (geth)
rollup failed, requiring a full restart of the service.
With this patch the `searcher` task of composer tries to immediately
restart a collector if it exited.

## Changes
- Handle a collector exiting (with or without error) by immediately
restarting it.
- Set the retry-count in the ethers `WsClient` (aliased to `Ws`) to 0:
the retry logic is opaque and has strange interaction with startup,
reconnect on websocket drop, and reconnect on subscription drop. We
better explicitly differentiate between connect-failure,
subscription-failure, and reconnect-retry.
- Change collector reporting its `is_connected` status only after both
the websocket client && the subscription are established.
- Move the`geth` server mock to `astria-test-utils` and put it behind
feature flags.

## Testing
- A unit test for the `reconnect_exited_collector` function is provided,
which triggers the reconnect-logic by returning a subscription error.
Warts:
- It would be better to test the TCP socket failing, but that would
create flaky tests because it would require binding the tcp socket to
the same port. An alternative would be to use tokio-turmoil, but ethers
does not allow configuring the TcpStream.
- It would also be better to test the whole `Searcher` instead of the
handler (because it feels like we are testing implementation details),
but it that would require mocking many more parts of it. A better
alternative would be to create a `CollectorManager` that abstracts the
reconnection + channel logic away from `Searcher`.
    - 
## Related Issues

closes https://github.com/astriaorg/astria/issues/504
closes https://github.com/astriaorg/astria/issues/283