refactor: remove signing and signature verification of data posted to DA (#538)
## Summary
remove signing and signature verification of data posted to DA so that
data can be relayed to DA by anyone, not just the block proposer.

## Background
We want validators to not have to also run a relayer. The sequencer
block data is already committed by the sequencer chain, so additional
signatures aren't necessary. Also, you can already see who
signed+submitted the data by reading the signature of the Celestia tx
that submitted it.

## Changes
- remove `SignedNamespaceData` and all mentions/usages
- update relayer to have a `disable_relay_all` option, which, when set
to true, relays only blocks signed by the given validator key. the
default value is `false`, so all blocks are relayed.
- update conductor to verify the block received from DA has the same
hash as the canonical block at that height returned by the sequencer

## Testing
unit tests

## Breaking Changelist
- data posted to DA is no longer in the form of `SignedNamespaceData`,
so deserializing previously-written data will fail.

## Related Issues

closes #523