feat(conductor): implement derivation and execution of optimism `DepositTransaction`s (#535)
## Summary

At a high level, this PR introduces the ability to bridge ETH from an
ETH-L1 to a rollup running on Astria. This occurs by updating the
consensus driver (ie conductor) to derive deposits from an ETH-L1 and
send them to a rollup node which is able to understand and execute these
deposits, thus minting the funds on the L2.

see also https://github.com/astriaorg/op-geth/pull/1 

## Background
We want to be able to use ETH as a gas token on rollups running on
Astria. As well, we want to see if we can use the op-stack as a bridging
solution for these rollups.

## Changes
- implement the `astria-optimism` crate which contains logic for reading
and submitting deposits to an `OptimismPortal` contract (the contract
deployed on L1 for op-stack rollups, which accepts ETH deposits)
- the `astria-optimism` crate also contains logic for converting
`TransactionDeposited` events (emitted when a deposit is made) into a
`DepositTransaction` which is a transaction type on the L2
- update conductor to have an optional optimism config, which, when
enabled, allows conductor to watch for `TransactionDeposited` on an ETH
L1 and convert and execute these as `DepositTransaction`s on the L2.
- TODO; we also need to figure out block timing and finality; this PR
will not sync deposit transactions properly if conductor restarts (it
will re-execute all of them, unless the
`initial_ethereum_l1_block_height` is updated), or deal with re-orgs on
the ETH chain. we need to figure out how to map sequencer/rollup blocks
to eth blocks in a deterministic way.

## Testing
astria-optimism: unit tests
conductor: unit test + manual testing, see
https://www.notion.so/astria-org/op-stack-on-astria-ba781ff4bc3244de84f91fef523dbb43?pvs=4

## Breaking Changelist
- none, as the optimism feature is disabled by default

## Related Issues
none

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>