fix(conductor): remove proposer check; use correct validator set in commit verification (#589)
## Summary
remove proposer check from block verifier as it's not needed, also the
logic was wrong anyways. the verification of the `Commit` is sufficient
to ensure the block was actually committed by the sequencer.

also update to use correct validator set in commit verification; we need
to use the height at `n-1` for a block at height `n`. see code comment
also. this makes sense as we have a validator set at genesis, which
votes on block 1; then, the set at block 1 votes on block 2, etc.

## Background
see above

## Changes
- remove proposer check from block verifier
- use correct validator set in commit verification

## Testing
these issues were found in devnet testing