fix(sequencer): update app state to latest committed before starting round (#673)
## Summary
update's the sequencer app state to latest committed before starting a
round. if the previous round's data was committed, this effectively does
nothing.

## Background
this is a bug; since we execute txs during prepare/process proposal,
this updates the app `StateDelta`, which was not reset in case of the
proposal failing to be committed. this results in a bad state.

## Changes
- update app state to latest committed before starting round during
prepare/process proposal

## Testing
should be tested on devnet