feat(conductor)!: forward firm blocks read from celestia in sequential order (#695)
## Summary
The celestia reader task only forwards blocks to the executor if their
heights are in sequential order.

## Background
As part of the changed contract the sequencer and celestia readers
forward their blocks to the executor only if they are in sequential
order of the sequencer heights recorded inside them.

## Changes
- Introduce a block cache that yields a new block with height H only if
it has previously yielded a block with height H-1
- Change the http-based celestia client to websocket
- Remove polling celestia for new blocks at fixed interval times
- Subscribe to new celestia headers, extracting the newest celestia
height
- Add a `ReconstructedBlockStream` that acts as a buffered queue
fetching celestia blobs at the the height obtained from the header
subscription

## Testing
No explicit tests. This must be tested end-to-end. Mock tests should be
added in a follow-up.

## Breaking Changelist
- must connect to celestia using websocket instead of http

## Related Issues
Part of https://github.com/astriaorg/astria/pull/691