feat(conductor): use http client to fetch celestia blobs (#704)
## Summary
This patch changed blob-fetching to use a http client. The websocket
subscription is still used for fetch the latest celestia header (and
hence height).

## Background
Reusing the same websocket client for jsonrpsee subscriptions and simple
methods requires some complex client pool + resetting logic. Using the
websocket client for the subscription only and http clients for fetching
blocks allows to simplify this significantly.

## Changes
- Add a http client to the `celestia::Reader`
- fetch celestia blobs using the http client instead of the websocket
client
- update the URI scheme of the celestia node API depending on whether
the websocket or http clients are to be constructed