refactor(conductor): interact with executor via handle (#711)
## Summary
Added an `executor::Handle` type to send blocks, read state.

## Background
The previous implementation required too much manual plumbing to set up
connections to the executor. A handle to a long running task is a more
elegant and idiomatic solution. It also allows to ensure at the
type-level that a `Handle::wait_for_init -> Handle<StateIsInit>` is
called prior to reading its state and sending data to it because all
methods to do so are only implemented on `Handle<StateIsInit>`.

## Changes
- Introduce an `executor::Handle<T>` type to interact with the
`Executor`.
- Move all send channels out of the executor and into the Handle
- Introduce `StateIsInit` and `StateNotInit` marker types to ensure the
executor state is initialized (which is the case when it read the
genesis info and commitment state off its rollup node)