feat(conductor): restrict firm<>soft commitment spread and add celestia search windows (#725)
## Summary
This patch restricts the maximum permitted spread between soft and firm
commitments to 16 heights and adds a search window to celestia heights
within which the next sequencer block must be found.

## Background
As part of the refactor it was found that the loop reading off of
sequencer was too hot, reading more sequencer blocks than could be
executed against the rollup node leading to conductor crashing with out
of memory. To fix this a, the soft-block channel from the sequencer
reader to the executor needed to be changed from an unlimited to a
limited channel.

With this rework it made sense to go all the way and also backpressure
on the celestia reader.

## Changes
- Executor tests if soft commitments are too far ahead of firm
commitments, and stops pulling soft blocks from sequencer-reader until
firm commitments have caught up (currently hard-coded at a difference of
16).
- The above leads to backpressure on sequencer reader, which stops
reading new sequencer heights until the channel is free
- Because the tendermint-rs JSONRPC subscription only allows for reading
full blocks the code is changed to not deserialize raw tx bytes and only
keep sequencer block heights. This is wasteful in network costs but
doesn't put pressure on conductor memory and will hopefully be addressed
with future tendermint-rs (or cometbft-rs) releases.
- The backpressure on the celestia-reader is indirect: it actively
monitors the commitment state reported by executor and only advances the
maximum celestia height that it will read based on the firm commitment
number.

## Testing
Conductor was redeployed several times with different configurations of
`FirmOnly`, `SoftOnly`, and `SoftAndFirm` commitment levels. It was
observed that it always reached the same commitment hash at a
non-trivial height (in this case 1961).