feat(conductor): set permitted commitment spread from rollup genesis (#743)
## Summary
Sets the permitted commitment spread between soft and firm commitment
heights to a number derived from the rollup genesis info.

## Background
The spread implemented in the refactoring
https://github.com/astriaorg/astria/pull/691 was set to a fixed number
of 16 which is very small and not dynamic. This patch takes the spread
as the `6 * celestia_block_variance`, with the variance set in the
rollup's genesis info. 6 is a heuristic taken as the upper limit of
expected sequencer heights per celestia height.

The channel to send soft blocks from sequencer reader to executor was
rewritten in terms of a semaphore that acts as a bound to an unbounded
mpsc channel. That is done because the bounded tokio mpsc channel (and
any other mpsc channel currently found in the ecosystem) does not allow
to have its capacity changed dynamically.

## Changes
- refactor the soft block channel in the executor to an unbounded
channel bounded by a semaphore
- set the channel's capacity from the celestia block variance times 6
- add permits to the channel if firm blocks have caught up with
sequencer blocks
- simplified rescheduling of rejected blocks due to backpressure by
leaning on async send methods (instead of putting them back in the
cache)
- hid the implementation details of sending blocks to the executor by
removing access to the channels in its handler

## Testing
Added unit tests to test the general flow of the channel (initializing,
adding permits, exhausting it).