feat(proto): add sequencer service proto (#701)
## Summary
Defines an API service to be implemented in sequencer which can be used
to decouple conductor & sequencer relayer from core transaction types.

## Background
Currently conductor & relayer utilize CometBFT apis to grab sequencer
data, and parse the transactions. This tightly couples them with
transaction formats, and increases computation cost in parsing each new
block. When implemented, this will enable the types to be decoupled in
conductor so we can change transaction formats without breaking rollups,
and streamline the code for relayer.

## Changes
- Adds a new service definitions

---------

Co-authored-by: noot <36753753+noot@users.noreply.github.com>
Co-authored-by: elizabeth <elizabethjbinks@gmail.com>
Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>