feat(conductor)!: provide explicit HTTP, Websocket Celestia RPC URLs (#780)
## Summary
Uses explicit configuration to set HTTP and Websocket URLs to interact
with a Celestia Node.

## Background
The previous way to set URLs to interact with the Celestia Node RPC was
confusing (with scheme? without scheme? with trailing slash? without?).
This patch removes all that complexity in favor of just setting strings.

## Changes
Changed environment variables to configure conductor (and associated
config options):
- Remove `ASTRIA_CONDUCTOR_CELESTIA_NODE_URL` env var
- Add `ASTRIA_CONDUCTOR_CELESTIA_NODE_HTTP_URL`
- Add `ASTRIA_CONDUCTOR_CELESTIA_NODE_WEBSOCKET_URL`

## Testing
This is a simplification that just shuffles values around. The presence
of the strings is enforced at compile via a type-state builder pattern.
This should be tested end-to-end.

## Breaking Changelist
- Need to update dev cluster config maps and templates to add/remove the
env vars mentioned in the Changes section.

## Related Issues
This is the better solution to
https://github.com/astriaorg/astria/pull/761, which should be closed.