chore(core): Update docs to reference astria-core instead of astria-proto (#777)
## Summary
astria-proto repo was renamed to astria-core in
https://github.com/astriaorg/astria/pull/644 . This PR addresses the
corresponding README changes to change `astria-proto` to `astria-core`

## Background
The
[https://github.com/astriaorg/astria/tools/protobuf-compiler/README.md](https://github.com/astriaorg/astria/tools/protobuf-compiler/README.md)
includes the path which points to where the generated code for the
protobufs resides. It points to a stale `astria-proto` link which can be
confusing to new developers looking at the codebase. The other README
changes come as a result of this observation.

## Changes
- Updates to READMEs

## Testing
No need of testing

## Metrics
No need of metrics

## Breaking Changelist
No breaking changes

## Related Issues


closes <!-- list any issues closed here -->