feat(sequencer): store sequencer blocks in the sequencer state (#787)
## Summary
generate and store `SequencerBlock`s in the sequencer app state for
usage in the sequencer API. now, the sequencer API does not need to call
cometbft and do heavy conversions, it just reads the `SequencerBlock`s
stored during app execution from state.

## Background
see above - less conversions and eventually we can remove the need for
relayer/conductor to call cometbft at all.

## Changes
- implement the `api_state_ext` module which writes a sequencer block to
state by splitting it into its different parts (header, rollup datas,
proofs) and writing those to disk separately. when fetching a filtered
sequencer block, only the relevant data needs to be fetched.
- allows fetching a block/data by block number or hash.
- implement the `SequencerBlockHeader` type which contains the cometbft
header and the relevant rollup data commitments, which really *should*
be part of the block header but we haven't forked cometbft. these fields
need to exist in the `SequencerBlock` type as they need to written to
disk, and the commitments need for the filtered sequencer block API. if
they're not stored, then fetching filtered block data will require the
entire block data to be fetched so that the commitments can be
reconstructed, which is not what we want.
- inside the app, implement `SequencerBlockBuilder` which stores the
header during `begin_block` and each tx during `deliver_tx`, and in
`end_block` uses these to construct and store the `SequencerBlock`.
note: this can be removed after #679 is implemented

## Testing
unit tests + i ran it with cometbft.
## Related Issues

closes #773

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>