fix(conductor): simplify mapping Celestia heights to Sequencer heights (#797)
## Summary
Simplifies mapping Celestia heights to Sequencer heights by adding the
source Celestia height to final type that contains the reconstructed
sequencer block information. Fixes an assertion (and panic) being
triggered by making it unecessary.

## Background
Conductor kept a map of `sequencer_height -> celestia_height` to track
which blobs contained which sequencer height. This implementation was
unnecessary and error prone, and ultimately out-of-sync with the rest of
the Celestia reader. It's select-loop contained an `assert!` macro
intended to tie assumptions about height-progress together and
ultimately revealed a bug.

Instead of keeping an extra map, the Celestia heights can be flattened
directly into the objects that are reconstructed from the blobs read off
of Celestia. This makes the assertion unnecessary and simplifies the
code.

## Changes
- Add a field `ReconstructedSequencerBlock::celestia_height` to track
the Celestia source heights.
- Remove the `SequencerHeightToCelestiaHeight` map
- Streamline the height tracking in the block stream.

## Testing
Should be tested end-to-end