feat(sequencer)!: include deposit data as part of rollup data (#802)
## Summary
include deposit data generated from `BridgeLockAction`s as part of
rollup data. this allows the relayer to write the deposit data to DA as
part of rollup data, and for rollups to receive deposits from the
conductor (once relayer/conductor are updated to use the sequencer API).

## Background
required for the bridge accounts implementation

## Changes
- implement `RollupData` enum type which consists of either
`SequencedData` or `Deposit`. this type is used to wrap either a
`SequenceAction`'s byte data or a `Deposit`.
- this type is now what's written to each rollup data field + the rollup
data merkle tree in the form of a proto-encoded `RollupData` . this
allows the rollup execution node to decode each piece of data it
receives from conductor into a `SequencedData` or `Deposit`. a rollup
can still choose to ignore deposits.
- update building/storing of `SequencerBlock` in the app to include
deposits generated by the block's execution.

## Testing
unit tests + i ran it with cometbft

## Breaking Changelist
- this changes the rollup data commitment by including deposits, as well
as formatting them as encoded `RollupData`. thus after this PR,
commitments made before this PR will not be valid.

## Related Issues

closes  #774