chore(composer): disambiguate chain-id (#791)
## Summary
Disambiguate uses of chain-id in core services, making sure it's clear
when we mean `rollup-id`, `rollup-name`, or `chain-id`

## Background
We previously used to have a lot of chain id overlap. Now chain id
(outside of the cli crate), should refer to only the cometbft chain id
of either the sequencer or another IBC compatible chain id. Instances
referring to rollups are now referenced wholly by `rollup-name` when it
is a string or `rollup-id` when it is a 32 bit byte array.

There is a previous stale PR
(https://github.com/astriaorg/astria/pull/633) that accomplished this,
but many change since then opted to go from scratch.

## Changes
- update references mostly in composer to `chain-id`