fix(conductor): serialize rollup IDs as strings so telemetry doesn't crash (#821)
## Summary
Ensures that rollup IDs are serialized as strings.

## Background
The serde derivation for `RollupId` was missing a `serde(transparent)`
attribute. This means that, for JSON, rollup IDs were not serialized has
`"<hex-of-32-bytes>"` but as `{ "inner": "<hex-of-32-bytes>" }`.

This in turn caused the JSON-serialization of some structs used for
reporting received blocks from sequencer to fail, causing a panic in the
tracing stack (as it assumes that `Display` formatting won't fail).

## Changes
- Add `serde(transparent)` attribute on `RollupId`
- Add snapshot tests to reporting types

## Testing
Added snapshot tests for types used in conductor tracing (that's where
we saw the failed display formatting and subsequent panic).