refactor(relayer): replace grab-bag constructor with config-like builder (#822)
## Summary
Explicitly populates a builder and constructs the relayer actor from it
instead of passing in the config.

## Background
Using a config as a grab-bag to initialize types is bad practice because
it's not immediately clear which option are relevant to which part of
the stack.

This patch adds a `relayer::Builder` type with crate-public fields which
are explicitly populated and then used to construct the final `Relayer`
object. This fulfills a similar need as the type-state builder pattern
used in conductor, but needs significantly less boilerplate. And because
there are no extra invariants that are supposed to be held up by the
builder-type itself this is all around nicer.

## Changes
- Add a `relayer::Builder` type
- Replace `Relayer::new` with `relayer::Builder::build`
- Remove unnecessary sources of `async` so that
`relayer::Builder::build` need not be `await`ed on:
    - instead of `celestia_rpc::Client` use `jsonrpsee::HttpClient`
- use `tonic::Endpoint::connect_lazy` to delay connecting until the
first attempt

## Testing
+ Most changes were shuffling code around and tests still pass.
+ Block fetching over gRPC still works, which means lazy client
initialization still works.
+ Added an Authorization middleware to the mocked celestia server to
ensure the changes to Celestia construction still work.

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>