chore(collector): Rename Collector to GethCollector (#792)
This PR renames Collector to GethCollector. It also renames the `Status`
structs to make it easier to distinguish.

As we are working towards adding different types of collectors to
Composer, it is better to re-name the existing Collector implementation
to `GethCollector` since it streams txs from the rollup node's geth
mempool.

- Rename `Collector` -> `GethCollector`
- Renames collectors field in `Searcher` struct to `geth_collectors`

Note: I chose not to name other fields of `Searcher` struct like
`collector_statuses` and `collector_tasks` since they can be generic on
different types of collectors and not just `GethCollectors`

Compiled the codebase and ran tests successfully.

https://github.com/astriaorg/astria/issues/795

closes <!-- list any issues closed here -->