fix: justfile updates to build dependencies (#867)
## Summary
This fixes three cases in the justfile where dependencies were not being
built as part of the relevant steps.

## Background
When running `just deploy [sequencer|rollup|dev-rollup]` the required
dependencies were not being built due to the commands being appended
with `| true`. This was intended to silence the output of the commands,
but resulted in the commands failing silently. I believe that's down to
`true` returning `0` well before the helm command has completed, causing
helm to receive a `SIGPIPE` signal and exiting with error.

## Changes
The `| true` instances were replaced with `> /dev/null`. It seems that
my editor has stripped trailing whitespace too.

## Testing
Manual testing only. There is a smoke test where these could be added,
but currently the smoke test explicitly calls the required `helm
dependency build` commands, so I'm not sure if they should be replaced
or left as is.