feat(composer): Add a gRPC collector (#784)
## Summary
We add a gRPC collector to the Searcher in the Composer module. The gRPC
collector exposes an API which allows end-users to send bytes of txs
along with their corresponding rollup ids.

## Background
Currently, the Composer has a Searcher component which spawns up various
rollup specific Collectors. These Collectors stream txs sent to the
rollup node on a websocket connection. Essentially we subscribe the
`getPendingTransactions` websocket stream of the rollup node. Here, we
assume that the rollup node is a Geth node.

We want to add an API to which users can submit a list of sequence
actions which will be bundled and sent to the sequencer. This will make
testing a lot easier and allow non-geth based rollups to integrate with
Astria composer easily.

We also want to keep the existing Composer flow of streaming txs from
the rollup nodes alongside with this API.

## Changes
- Added a gRPC server in the `Composer`.
- Add a new configuration `grpc_addr`. This is the address to which the
gRPC collector is binded to. We bind to this address in
`Composer::from_config`
- We add the protobufs for this gRPC server in the `proto/composerapis`
directory. We specify the protos for the
`SubmitRollupTransactionsRequest` which is the request sent to the gRPC
server and the `SubmitRollupTransactions` rpc which handles validating
the payload received in `SubmitSequenceActionsRequest` and its passing
to the `Executor`
- Added the tonic_health healthcheck service to the gRPC server.

## Testing
Added integration tests where we submit txs to the gRPC collector and
verify if it was included in the Sequencer block with a mock sequencer.

## Notes
1. There is a field called `rollups` in cfg which is basically a map
from rollup to the websocket url to be used by the GethCollector. We
could better call it `rollup_by_id` but thats a breaking change which I
think we can reserve for future PR if required since it could break any
existing integration testing systems.
2. We do not track the status of the gRPC collector in the
composer_status. This is because the gRPC collector essentially just
starts up a gRPC server and if there is a failure during the
startup/running of the gRPC server, it is a pretty critical issue.

## Issue
https://github.com/astriaorg/astria/issues/795

closes <!-- list any issues closed here -->