feat(sequencer-relayer): provide a shutdown controller (#889)
## Summary
Adds a shutdown handle for the sequencer-relayer.

## Background
We want the ability to invoke the shutdown sequence from `main` and from
tests.

## Changes
A new RAII object `ShutdownHandle` has been added. This cancels the
wrapped `CancellationToken` when the `ShutdownHandle` is dropped or when
its `shutdown` method is called.

`SIGTERM` handling has been moved to main.rs; on receiving a `SIGTERM`,
`shutdown` is called on the shutdown handle associated with the relayer.

The relayer's main `select!` loop was simplified to only contain the two
tasks.

## Testing
The `TestSequencerRelayer` was updated to make use of the new shutdown
controller. Also manually tested by sending a `SIGTERM` to a running
sequencer-relayer.

## Related Issues
Closes #882