feat: add gRPC mocking library (#872)
## Summary
Adds a library for mocking gRPC servers heavily inspired by wiremock-rs.

## Background
conductor, sequencer-relayer, and soon composer are interacting with
gRPC servers. In blackbox tests these services use ad-hoc
implementations of gRPC servers with manually crafted internal state to
return desired responses. This is very tedious and error prone work, and
moving forward not sustainable.

Specifically conductor is lacking blackbox tests entirely. Implementing
blackbox tests for it is the motivation for `astria-grpc-mock`.

This library is heavily inspired by
https://github.com/LukeMathWalker/wiremock-rs. Deviating from it,
`astria-grpc-mock::MockServer` does not stand in as a full server in its
own right, but is run next to a `tonic::Server`. Implementors of a
tonic-generated service then delegate individual trait methods to
`MockServer::handle_request`. See the tests in `astria-grpc-mock-test`
how this is done.

The other big difference is that we are heavily leaning on
`std::any::Any` to erase the types of concrete gRPC requests and
responses, i.e. for functions of the form `rpc: T -> U` both `T` and `U`
are erased. Because a mock author needs to provide an RPC respsonse `V`,
`astria-grpc-mock` panics if `U` and `V` don't match, failing the test.
This check happens if the `MockServer` goes out of scope, if a
`MockGuard` goes out of scope, or in the
`MockGuard::wait_until_satisfied` future (different from `wiremock-rs`,
where `MockGuard::wait_until_satisfied` cannot panic).

Documentation and ergonomics of this library will be improved in
follow-up PRs (an idea is to use a proc macro to automatically generate
a desired mock implementation for a server, but the details around how
to this are complex).

**Note to the reviewer:** The best way to review this library is to see
how it is used in `astria-grpc-mock-test` and move backward from there.

## Changes
Added the following libraries:
- `astria-grpc-mock` to write grpc server mocks
- `astria-grpc-mock-test` to test the public interface of
`astria-grpc-mock`
- `astria-grpc-mock-codegen` to generate Rust bindings with pbjson
`Serialize` impls for the google `grpc.health.v1` package (this taken as
is from `tonic-health`)

## Testing
The `astria-grpc-mock-test` crate tests the `astria-grpc-mock` crate. No
other tests are necessary as this crate does not show up in release
builds of any service.