fix(sequencer)!: fix state setup to be consistent before transaction execution (#945)
## Summary
we had an issue with syncing blocks with a full node due to the state
not being set up the same for tx execution in the proposal phase vs not
in the proposal phase. previously, this did not cause issues as txs were
not dependent on the state set up when beginning the block, but with the
introduction of ibc actions, they now are.

## Background
see above

## Changes
- refactor the sequencer app to have a fn `pre_execute_transactions`
which must be called before transaction execution in any phase. this
function is then called in `prepare/process_proposal`, which sets up the
state to be the same as when a block is executed via only
`finalize_block`.
- rename some functions for clarity (deliver_tx -> execute_transaction,
execute_block_data -> execute_transactions_before_finalization)

## Testing
unit tests + we should run it on devnet

## Breaking Changelist
- state is now set up differently than before during prepare/process
proposal