fix(sequencer): don't store execution result of failed tx (#992)
## Summary
this was causing a cometbft consensus failure - we should only return tx
results for txs included in the block, which are only txs which pass
execution. cometbft requires that the number of tx results returned
equals the number of txs in the block (would be bad otherwise)

## Changes
- don't store execution result of failed tx in the proposal execution
phase, fixes consensus failure as it now returns the correct number of
execution results.

## Testing
ran with cometbft v0.38.6