feat(conductor)!: fetch blocks pending finalization  (#980)
## Summary

Adds the ability for Conductor to sync its connected
rollup if firm commitments trail soft commitments.

## Background
Conductor uses the
`astria.execution.v1alpha2.BatchGetBlocks` API to
retrieve all blocks in the range `firm+1..=soft`
from the rollup node. For example, if the firm
number is 3 and soft number is 10, then all
soft-committed blocks 4 through 10 (inclusive) are
fetched.

Previously conductor would fail try to re-execute
the firm blocks as it didn't have their
corresponding rollup-hashes, causing the rollup
node to reject the execution.

## Changes

1. Conductor now matches rollup block numbers
rollup blocks (previously, rollup block hashes,
which it doesn't have on a fresh start).
2. Conductor uses these block numbers to update
the commitment state upon receiving a pending
firm block.
3. Reworked the channel tracking the rollup state
in the executor task: since Conductor must now be
able to in both directions between Sequencer
heights and Rollup numbers, updating the tracked
commitment state is now a fallible operation. This
required an abstraction on top of the
watch-channel containing the state.
4. changed how the next expected Sequencer height
is calculated from the Rollup number (for both
firm and soft commitments). It is now
sequencer_genesis_block_height + block_number + 1
(previously without + 1, which was confusing).

## Testing

1. Added 2 blackbox tests for soft-and-only mode,
one for testing the pending-block fetch.
2. Removed the executor tests as they are now covered
by the blackbox tests.

## Breaking Changelist

1. The calculation of the next expected Sequencer
height is now one more than previous. This is a
breaking change in the sense that rollup + conductor
need to started with the same view of
rollup number <> sequencer height in mind. A rollup
that was started with an older conductor will not work
with this conductor.

## Related Issues
Link any issues that are related, prefer full github links.

closes https://github.com/astriaorg/astria/issues/837