fix(conductor): only execute firm blocks if firm and soft block numbers match (#1021)
## Summary
Ensures that conductor only executes firm blocks if running in
soft-and-firm mode, and if the rollup soft and firm block numbers are
equal.

## Background
Previously, Conductor used the presence of a block in its cache [1] as
the condition for whether it should execute a firm block against a
rollup. Blocks missing from this cache should never be observed in
practice, but there is a better (less confusing, more maintainable way)
of expressing the same condition:

When running in soft-and-firm mode, Conductor:
1. must only execute a block if the rollup's `firm number == soft
number`, and then use the rollup response to advance the firm commitment
state.
2. must otherwise only advance (without prior execution) the rollup's
firm commitment state using the block pending finalization for the
corresponding block number stored in its cache.

1: the cache of blocks pending finalization; only relevant in
soft-and-firm mode.

## Changes
- Tightened the condition for when Conductor executes a firm bock
against the rollup when running in soft-and-firm mode (only execute if
soft number equals firm number)
- Makes the absence of a block in its cache a hard error when not
executing but only advancing the commitment state
- Tries to fetch the missing block pending finalization from the rollup
as a last ditch effort before bailing

## Testing
Added tests for determining if a firm block should be executed. Because
the case where Conductor tries refetching the missing block is a
fallback for a faulty edge case that should never be observed in
practice this cannot be tested using blackbox tests.

## Related Issues

closes https://github.com/astriaorg/astria/issues/1011