feat(sequencer-relayer)!: provide filter for rollups (#1001)
## Summary
A new configuration env var has been added to allow filtering data
submission to a subset of rollups.

## Background
This allows more flexibility in what is actually submitted to the
Celestia chain.

## Changes
A new env var was added, mapped to a new field
`Config::rollup_id_filter`. This represents a list of rollups whose
transactions should be included in blobs submitted to Celestia. Rollups
not in the list do not have their data submitted. However, if the list
is empty, no filtering is applied; all rollups have their data
submitted.

The collection of filtered rollup IDs is passed down to the
`write::conversion::convert` function where sequencer blocks are
converted to blobs. The blobs are filtered inside this function, partly
since this is relatively early in the flow, but also so that the
`ConversionInfo` struct can carry information about blobs being
excluded. This data is logged at the start of `write::submit_blobs` in
an info-level log message. It might be worthwhile adding metrics around
the number of excluded rollups per submission, but I didn't see a clear
benefit to that and it can be added in a follow-up PR if required.

## Testing
- Added unit tests covering parsing of the config string to a collection
of `RollupId`s.
- ~Added a black box test where a sequencer block is constructed with
several transactions, half from included rollups and half from excluded
ones. Existing black box tests ensure that when the filter is empty, no
filtering is done.~ These all got removed in #963, and should be
restored as part of #1008.

## Breaking Changelist
- Added `ASTRIA_SEQUENCER_RELAYER_ONLY_INCLUDE_ROLLUPS` which can be
empty or can be a comma-separated list.