refactor(sequencer): store fees for actions in app state (#1017)
## Summary
store fees for actions in app state, fees are of the form `base_fee +
byte_cost_multiplier * byte_size`.

for txs that don't write to DA, they have an implicit `byte_size` of
zero, so they only have a base fee.

for txs that write to DA, they have a `byte_cost_multiplier` which is
multiplied by the number of bytes written to DA for fee calculation.

## Background
 this is so eventually we can change fees via a tx.

## Changes
- store fees for actions in app state

## Testing
unit tests

## Related Issues
part of  #1003