fix(charts): deployment and celestia-node fix (#1038)
## Summary
Fixes the celestia-node chart and adds options for setting trusted
header for sync startup. Updates the deployment of charts so that old
files that were deleted will be removed, and updates the just files to
be an import so it can be utilized in the charts as well.

## Background
Celestia node chart would not run, and needs to sync from head, and the
rollup chart ended up incorrect due to the deployment leaving an old
deployment file. This version of the action deletes the old files in the
directory on each push.