feat(sequencer-relayer): add metric recording highest sequencer block submitted (#1040)
## Summary
Added a metric to the relayer to give visibility of the current highest
sequencer block stored on Celestia.

## Background
It will be useful to avoid having to search the logs for this info.

## Changes
Added the new metric (see below). Also dropped the log level of a couple
of somewhat noisy `debug!` log lines.

## Testing
No formal tests added. Viewed the metrics of a running relayer to ensure
the new metric was as expected.

## Metrics
- Added `astria_sequencer_relayer_sequencer_submission_height`.