feat(sequencer)!: implement bridge unlock action and derestrict transfers   (#1034)
## Summary
Implements bridge accounts' `BridgeUnlock` action and allows bridge
accounts to transfer assets like normal.

## Background

We need a `BridgeUnlock` action to help with tracking/proving that
withdraws for rollup specific events happened. This PR implements the
ground work for that by adding a new action with a `memo:bytes` field
that is currently unused.

We also decided that it's unnecessary to have restrictions on the bridge
account in terms of normal transfers in/out of the account. The bridge
account is already a privileged actor and having the restrictions on
transfer would make normal account maintenance (like gas funding) messy.

## Changes
- implement `BridgeUnlockAction` for transferring out a bridge's bridged
asset
- derestricted the transfer function to let bridge account also act like
normal accounts

## Testing
unit tests

## Breaking Changelist
- one new action is added: `BridgeUnlock`
- bridge account now can transfer assets in/out like normal sequencer
accounts

## Related Issues
closes #983