chore(sequencer-relayer): change compression ratio calculation (#1075)
## Summary
Inverts the compression ratio calculation.

## Background
It's normal to express a compression ratio as uncompressed:compressed,
but were reporting it as the inverse of this in logs and metrics
(despite the metric's description claiming it was
uncompressed:compressed).

## Changes
As per summary.

## Testing
Added a unit test.

## Metrics
No new metric added, but the value of
`astria_sequencer_relayer_compression_ratio_for_astria_block` should now
generally be > 1 rather than < 1.