feat(composer)! avoid holding private key in env var (#1074)
## Summary
This PR changes the configuration env vars for composer to not hold the
contents of the private key for signing sequencer transactions.

## Background
It's undesirable to have the contents of a private key in an env var.

## Changes
Changed the existing env var to hold a path to a hex-encoded private key
file.

## Testing
Modified existing tests, and ran smoke test using local build of
composer.

## Breaking Changelist
- Removed `ASTRIA_COMPOSER_PRIVATE_KEY`
- Added `ASTRIA_COMPOSER_PRIVATE_KEY_FILE`

## Related Issues
Closes #993.