refactor(sequencer): remove mint module (#1134)
## Summary
remove the mint module and associated types.

## Background
we never used it, and not something we want in production anyways.

## Changes
- remove the mint module and associated types.

## Testing
n/a, code was removed


## Breaking changes
this removes the mint action, which is technically a breaking change,
however it was gated under a feature flag in sequencer which was never
enabled on any network afaik. so I don't think this change affects any
running network.