fix(withdrawer): update AstriaWithdrawer to check that withdrawal value is sufficient (#1148)
## Summary
update AstriaWithdrawer to check that withdrawal value is sufficient.
for example, say a bridged asset is multiplied by 10^9 upon bridging to
the rollup from the sequencer. you should not be able to withdraw some
value <10^9, as this is not a valid asset amount on the sequencer, as
it's smaller than the smallest denomination on the sequencer. this
updates the contract to perform that check. the alternative is to update
the withdrawer not to send sequencer withdrawals with value=0, but this
method is better for ux as it won't allow a user to make an invalid
withdrawal in the first place.

## Background
found while debugging

## Changes
- update AstriaWithdrawer to check that withdrawal value is sufficient,
ie. value >= `10^assetWithdrawalDecimals`

## Testing
unit test