feat(sequencer-relayer)!: add chain IDs for sequencer and Celestia to config env vars (#1063)
## Summary
Added two new configuration options specifying the expected chain IDs of
the sequencer and Celestia networks.

## Background
This is a guard rail to help avoid an operator accidentally connecting
the relayer to the wrong network.

## Changes
The configured values are used in the initialization phase, where the
relevant clients query the appropriate servers for their chain IDs.

The Celestia client was already retrieving this info prior to this PR.
Now it returns an error if the chain ID doesn't match, and this
particular error variant causes the endless retry loop (and ultimately
the process) to exit.

The sequencer client was not previously requesting this info, so it now
does this in a somewhat similar manner to the Celestia client. Likewise,
a chain ID mismatch there causes the process to exit.

## Testing
A blackbox test for a mismatch in each of the chain IDs was added.

## Breaking Changelist
- added `ASTRIA_SEQUENCER_RELAYER_SEQUENCER_CHAIN_ID`
- added `ASTRIA_SEQUENCER_RELAYER_CELESTIA_CHAIN_ID`

## Related Issues

Part of #986.
Closes #978.