feat(withdrawer): bridged ERC20 token withdrawals (#1149)
## Summary
implement withdrawals of ERC20 tokens that are of type
`AstriaBridgeableERC20` (see implemented contract).

## Background
we want to be able to withdraw ERC20 tokens that are bridged to a
rollup.

## Changes
- implement `AstriaBridgeableERC20` which is a standard `ERC20` contract
with additional functionality for minting (not used by the withdrawer,
implemented/tested here
https://github.com/astriaorg/astria-geth/pull/20) as well as
functionality for withdrawing
- implement `IAstriaWithdrawer` which is implemented by both
`AstriaWithdrawer` and `AstriaMintableERC20`.
- the withdrawer now interacts with a `IAstriaWithdrawer`. both native
assets and ERC20 withdrawals have the same event signatures, so no
additional code was needed for the withdrawer itself.
- to use the withdrawer with an `AstriaBridgeableERC20`, the
`ASTRIA_BRIDGE_WITHDRAWER_ETHEREUM_CONTRACT_ADDRESS` is set to some
`AstriaMintableERC20`, and
`ASTRIA_BRIDGE_WITHDRAWER_ROLLUP_ASSET_DENOMINATION` is set to the
rollup asset's denomination as represented on the sequencer. for
example, if the asset is represented on the sequencer is
`transfer/channel-1/usdc`, that is the rollup asset denomination. the
`name/symbol` of the ERC20 contract are not relevant.
- also update the build script to write the generated abigen contract
bindings to files, this is easier for debugging and unit testing.

## Testing
unit tests

## Related Issues

closes #924