chore: register all metrics during startup (#1144)
## Summary
Metrics in all five binaries are registered early in runtime flow.

## Background
For a given binary, we want the metrics http endpoint to always yield
data for every metric relevant to that binary, even if some of them have
not been modified since startup. The `metrics::describe_` macros
(currently called during startup) don't register the provided metrics -
instead individual metrics are only registered on the first execution of
the relevant `counter!`, `gauge!` or `histogram!` macro.

Further, the descriptions were never being applied, since in
`astria-telemetry` the call to these macros was made before the metrics
builder was installed.

## Changes
Similar changes were made to `astria-bridge-withdrawer`,
`astria-composer`, `astria-conductor`, `astria-sequencer` and
`astria-sequencer-relayer`:
- `metrics_init` module renamed to `metrics`, and used to hold a
`Metrics` struct containing the relevant metrics objects.
- a static ref to a `Metrics` struct is created early in the flow (but
after the call to telemetry's `Config::try_init` where the metrics
builder installs the global registry) and passed to whichever
objects/functions require it. This approach will ultimately allow us to
choose whether to use the global metrics registry or a local one, making
the metrics much more testable in the future.
- removed telemetry `Config`'s `metric_buckets` since it wasn't really
useful (it applies the bucket limits to all histograms, regardless of
type), and `register_metrics` since it's also no longer used.
- changed the units of several `_latency` histograms from ms to seconds,
since they were actually recording seconds. This is a non-issue though,
as the units are ignored in Prometheus.
- added a description for
`astria_conductor_sequencer_blocks_metadata_verified_per_celestia_fetch`
as this was missing.

## Testing
Manually tested that all metrics for a couple of the binaries are
available in full as soon as the binary is started. Programmatic testing
can be added in follow-up PRs.

## Metrics
- Changed the type of `astria_composer_transactions_dropped_too_large`
from gauge to counter. We previously had `describe_counter!` to describe
it and it was only ever being incremented by 1, but it was accessed via
`gauge!` which seemed like a mistake, so I changed its type to
`Counter`.
- `astria_composer_transactions_received` and
`astria_composer_transactions_dropped` each used to have two labels; one
to ID the rollup and the other specifying collector type. For the geth
collector, the `rollup_id` label used the rollup's chain name, whereas
for the grpc collector it used the rollup ID for the `rollup_id` label.
I changed these two metrics to always provide three labels;
`collector_type`, `rollup_id` and `rollup_chain_name`.

## Related Issues
Closes #1027 and #1100.