feat(sequencer): add ttl and invalid cache to app mempool (#1138)
## Summary
This PR adds to the App's mempool a way to signal to CometBFT when
transactions should be removed from the CometBFT mempool.

## Background
The CometBFT and App mempools currently get transactions that fail to
execute stuck in them. This is because when a transaction fails to
execute in `prepare_proposal()`, it doesn't get removed from the
CometBFT mempool. CometBFT only clears out transactions if they either
fail `handleCheckTx()` or are included in a block. Because of this,
CometBFT will re-add the failed transaction to the App mempool during
its `handleCheckTx()` maintenance, which will cause it to be fed to
`prepare_proposal()` again.

We also need a way for full nodes to clear out these invalid
transactions. Since these nodes don't run `prepare_proposal()`, we need
an additional way, like a tx TTL, to signal when transactions can be
dropped.

## Changes
- Added a transaction cache to the App's mempool which signals to
CometBFT when a transaction should be removed.
- Tracks in the App's mempool when a transaction is first seen.
- Transactions that fail in `prepare_proposal()` get added to the App's
removal cache.
- Transactions that are older than 10 minutes are added to the App's
removal cache.

## Testing
Unit tests and local testing of invalid transactions. 

## Metrics
Added `CHECK_TX_REMOVED_FAILED_EXECUTION` counter to the sequencer's
metrics.
Added `CHECK_TX_REMOVED_EXPIRED` counter to the sequencer's metrics.

## Related
Closes #979