fix(sequencer-relayer): avoid hanging while waiting for submitter task to return (#1206)
## Summary
This fixes a bug whereby the relayer hangs during shutdown due to the
submitter task not being instructed to shut down.

## Background
Bugfix.

## Changes
- Rather than cloning a single cancellation token, we now have the
following hierarchy:
- `SequencerRelayer::new` constructs a `ShutdownHandle` which wraps the
top-level cancellation token. This handle is held as a member variable
of `SequencerRelayer` and is used to cancel all relayer child tasks if
the API task exits. A clone of the `ShutdownHandle` is also held in
`main` and is used to cancel all tasks if a `SIGTERM` is received.
- A child of `ShutdownHandle`'s token is provided to the relayer task,
named `relayer_shutdown_token`. It is never actively cancelled - only
ever via the parent token.
- A child of `relayer_shutdown_token` is provided to the submitter task
and also held as a member of `Relayer`. It is actively cancelled after
the relayer's main `select!` loop has exited if the submitter task is
still running.

## Testing
Manual test only. I could see no way to artificially cause the relayer's
`select!` loop to exit. We could potentially try dropping in some
`#[cfg(test)]` gated code in the loop to force this, but that would also
mean we couldn't use the blackbox tests for it.

## Related Issues
Closes #1095.