feat(sequencer-utils): generate example genesis state (#1224)
## Summary
Provides a `generate-genesis-state` subcommand to sequencer-utils.

## Background
We had a `test-genesis.json` file committed to the repo that was not
checked to be in-sync with the most recent version of sequencer's actual
`GenesisState`. The subcommand `generate-genesis-state` removes the need
to keep the file in sync because one can always generate one that
matches the current format.

## Changes
- remove `astria-sequencer/test-genesis-app-state.json`
- move sequencer's `GenesisState` to the new `astria-core::sequencer`
submodule. Because this config is effectively a public API, it is
justified to be outside of the `astria-sequencer` binary crate.

## Testing
It is a very thin binary which can be run by hand. The specific format
of `GenesisState` not changing is ensured by a snapshot test that was
moved from `astria-sequencer` to `astria-core`.

## Related Issues
Closes https://github.com/astriaorg/astria/issues/1212

---------

Co-authored-by: noot <36753753+noot@users.noreply.github.com>