refactor(bridge-withdrawer): move generated contract bindings to crate (#1237)
## Summary
Moves the generated Rust bindings for Astria's bridge smart contracts to
a free standing crate.

## Background
Breaking out the bindings outside the `astria-bridge-withdrawer` binary
crate is needed to share code with a CLI tool to manually initiate
withdrawals.

This patch also uses the opportunity and makes the process of
re-generating bindings (and committing them to the repo) more similar to
how protobuf-bindings are generated: `tools/solidity-compiler` updates
the git submodule in `astria-bridge-contracts` and then regenerates all
bindings. The bindings are placed in
`astria-bridge-contracts/src/generated` and commited to the repository.

## Changes
- Create `crates/astria-bridge-contracts` crate and move the submodule
of the same name there
- Create `tools/solidity-compiler` tool and use it to generate rust
bindings for the bridge contracts
- Update `astria-bridge-withdrawer` to use `astria-bridge-contracts` as
a dependency
- Add a job `solidity-contracts-compiled` to ensure that the contracts
are up-to-date.

## Testing
Apart from some types being renamed, nothing has changed.