fix(sequencer)!: store native asset ibc->trace mapping in init_chain (#1242)
## Summary
we need to store the native asset ibc to "trace" mapping in the state,
otherwise queries for the native asset using the ID will fail. for
example `get_bridge_account_info` where the asset is the native asset
fails right now

## Changes
-  store native asset ibc->trace mapping in `init_chain`
- also enforce that the native asset is is "trace" form, as otherwise,
we won't be able to map from ibc->trace form for the asset as we don't
know the trace form.

## Breaking changes
- this is unfortunately breaking since the ibc->trace mapping is stored
in app state.