fix(conductor): don't panic during panic (#1252)
## Summary
Avoid panicking if a test is already panicking.

## Background
The Conductor test environment is checking if it shut down cleanly for
every test. If not, it's panicking to signal that the test should fail.

However, if a test was already failing (panicking) due to another bad
condition, this lead to the test being immediately aborted without
printing any useful diagnostics.

This patch fixes this by only panicking if the test is not already
panicking, and issuing a debug level message otherwise.

## Changes
- Change the `Drop` impl for the conductor test environment to only
panic if the test is not already panicking.

## Testing
This is to fix bad tests not providing useful information, which was
encountered while writing another PR.