chore: update `bytes` and `ics23` crates (#1279)
## Summary
Updated the `bytes` and `ics23` crates.

## Background
The version of `bytes` we currently use has been yanked due to [a
bug](https://github.com/tokio-rs/bytes/pull/718). Running `cargo audit`
exposed this as an allowed warning.

We also got a report of an issue with the `ics23` crate and a
recommendation to update to v0.11.3.

## Changes
- Ran `cargo update -p bytes` and `cargo update -p ics23`.

## Testing
No additional tests required.