refactor(core, proto)!: define bridge memos in proto (#1285)
## Summary
Defines the bridge unlock, ics20 withdrawal, and ics20 deposit memos as
protobuf.

## Background
The various memos were read inside sequencer to inform the behaviour of
the protocol, but were not defined in the canonical
`astria.protocol.v1alpha1` protobufs. This patch defines the memo types
in a new `astria.protocol.v1alpha1.memos` so that there is a single
source of truth for all protocol-relevant objects. The memos are still
only ever serialiazed as plaintext `JSON` and not as protobuf.

## Changes
* Define the following types:
  * `astria.protocol.v1alpha1.memos.BridgeUnlock`
  * `astria.protocol.v1alpha1.memos.Ics20WithdrawalFromRollup`
  * `astria.protocol.v1alpha1.memos.Ics20TransferDeposit`
* Remove the `astria_core::bridge` module
* Update `astria_bridge_contracts`, `astria_bridge_withdrawer`, and
`astria_sequencer` to operate in terms of these types
* Update all memo fields containing bytes to be of type `string` so that
the rollup-native formatting can be used (to make it easier to identify
e.g. transactions and avoid the use of base64 for non-Astria byte
slices)
* Various field names have been updated to clarify that their source is
the rollup (e.g. `rollup_transaction_hash`, `rollup_return_address`).

## Testing
The changes in this code rely on smoke tests still working. Because the
IBC tests require the latest (rust) astria-cli, the IBC tests were
reworked.

## Breaking Changelist
* This is a protocl breaking change: the shape of the memo JSON objects
has changed (object keys were renamed, the encoding of byte fields no
longer requires base64).

## Related Issues
Closes #1286

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>