chore(core): remove ed25519_consensus from public API (#1277)
## Summary
Ensures that ed25519_consensus is not part of the public `astria-core`
API.

## Background
The definition `struct astria_core::crypto::Error(#[from]
ed25519_consensus::Error)` leaks `ed25519_consensus::Error` into the
public API of `astria-core`. This patch removes the `#[from]` attribute
and updates a few lines by explicit constructors, `map_err(Error)`.

## Changes
* Remove the `From<ed25519_consensus::Error> for crypto::Error` impl,
removing `ed25519_consensus` from `astria-core`'s public API.

## Breaking Changelist
This is a breaking change in `astria-core` but not in any of our
services. Because we don't yet publish `astria-core`, this is not a
breaking change at this moment but would be breaking in the future.

## Related Issues
Closes https://github.com/astriaorg/astria/issues/1221