refactor(ci): execute ics20 test from statefulset celestia container (#1302)
## Summary
Updates the `charts/celestia-local` to spin up `celestia-appd` in a
stateful set. Executes the ics20 transfer inside the celestia-app
container inside the set.

## Background
https://github.com/astriaorg/astria/pull/1285 updates the ibc ics20
transfer test to require `celestia-appd` and an initialized celestia
keystore to be present on the machine running the test. This is
undesirable because it pollutes the host environment and creates
potential for collisions. Turning `celestia-local` from a kubernetes
deployment to a stateful set allows executing the ics20 transfer from
inside the container.

## Changes
- Rename `charts/celestia-local/templates/deployment.yaml ->
charts/celestia-local/templates/statefulsets.yaml`
- Changes `kind: Deployment -> kind: StatefulSet` in that file
- Update recipes in `charts/deploy.just` to use `kubectl rollout status`
instead of `kubectl wait` (the latter does not seem to work with
stateful sets)
- Use `kubectl exec` against the celestia-app container in the stateful
set to initialize the transfer instead of a locally present
`celestia-appd`

## Testing
This is an update to the ibc ics20 smoke test flow. The test still
passes with the expected result.

## Related Issues
Closes #1296
Closes #1295