fix(proto): fix import name mismatch (#1380)
## Summary
Updates the generated path names to match the actual proto dependency
name.

## Background
Prost makes assumptions on how modules will be imported, where the proto
and rust generated code modules match, but does not auto generate the
module. We had a mismatch of singular -> plural in two places this
resolves that.

## Changes
- updates generated rust `account` and `transaction` modules to
`accounts` and `transactions`

## Testing
test suite