fix(sequencer)!: fix block fees (#1343)
## Summary
Corrected block fee reporting so that proposer is paid the correct fees.

## Background
Previously, only `TransferAction` and `SequenceAction` called
`get_and_increase_block_fees()`, the method for increasing the running
fees for the block. In `end_block`, this was used for calculating how
much to increase the proposer's balance meaning that they were only paid
the fees for `Transfer` and `Sequence` actions.

## Changes
- Implemented `get_and_increase_block_fees()` in all actions which
charge fees.

## Testing
Added unit tests for each action, with the exception of
`Ics20Withdrawal`. This test should be implemented, but will require a
more in-depth configuration.

## Breaking Changelist
- App hash changed, snapshot has been regenerated.

## Related Issues
closes #1333

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>
Co-authored-by: Fraser Hutchison <190532+Fraser999@users.noreply.github.com>
Co-authored-by: noot <36753753+noot@users.noreply.github.com>