refactor(core, proto)!: define app genesis state in proto (#1346)
## Summary
Defines `astria.protocol.genesis.v1alpha1.GenesisAppState`, replacing
the Rust-as-JSON definition.

## Background
All protocol relevant Astria types are supposed to be defined in
Astria's protobuf spec. With
https://github.com/astriaorg/astria/pull/1285 having redefined the memos
as protobuf message, this patch migrates the last type to protobuf spec.

## Changes
- Define `astria.protocol.genesis.v1alpha1.GenesisAppState` and related
protobuf messages
- Remove the `astria-core::sequencer::GenesisState` module
- Update Sequencer in terms of the protobuf type
- Update all charts and snapshots (especially the genesis template)

## Testing
All tests have been updated to use the new types, including snapshot
tests. The genesis state is only read at `init-chain` and does not
affect the evaluation of the state machine. Hence no tests should change
as long as the same data is passed in (which is reflected by
`tests-breaking-changes` leading to the same hash).

## Breaking Changelist
This is a breaking change because an old sequencer would not understand
a new (json-formatted) genesis app state and vise versa.

## Related issues
Closes https://github.com/astriaorg/astria/issues/1347