Chore(Charts): seq-faucet bech32m chart update (#1301)
## Summary
Updates `sequencer-faucet` chart to support the changes made to the
sequencer faucet image for bech32m compatibility.

## Background
With the introduction of bech32m prefixed addresses, an update to the
sequencer faucet was required. This PR updates the `sequencer-faucet`
helm chart to use the new faucet.

## Changes
* ADDED `SEQUENCER_FAUCET_MINUTE` - rate limit on requests, default `1
day`
* ADDED `SEQUENCER_ADDRESS_PREFIX` - sequencer prefix address, default
`astria`
* ADDED `SEQUENCER_NATIVE_DENOM` - asset to transfer upon request,
default `nria`

## Testing
Tested locally against the latest kubernetes cluster