fix(test): use correct conditions to dermine smoke test success (#1395)
## Summary
Fixes the conditionals that determine whether a smoke test passes.

## Background
Many smoke tests incremented a fetch-counter as long as it was `counter
< max`, but the test was for `counter > max`, which was never reached,
and hence the test passed succesfully even though it actually failed.
This was easily overlooked because so far the tests make use of the
posix-compatible `[ $counter -lt $max_counter ]` and `[ $counter -gt
$max_counter]`. But since all tests require bash this patch changes it
to use `(( ))`.

## Changes
- Change all smoke test scripts to use the the bashism `(( $counter <
$max_counter ))` for incrementing the counter, and `(( $counter >=
$max_counter ))` to check for failure.

## Testing
This correctly fails tests if the conditions don't apply (tested offline
and verified in CI).

## Related Issues
Partially addresses https://github.com/astriaorg/astria/issues/1392.
This patch will fail CI until after
https://github.com/astriaorg/astria/pull/1393 is merged.