fix(sequencer): update mempool benchmarks (ENG-733) (#1385)
## Summary
Recent changes to the mempool broke the benchmarks.  This PR fixes that.

## Background
Specifically, the benchmarks tests were failing in two ways: the limit
on number of parked txs per account was causing the addition of txs to
fail, and the pending nonce was returning `None` in the tests.

This was all due to the current account nonce being provided as 0, but
the txs started from nonce 1, meaning they weren't eligible for adding
to the pending queue.

## Changes
- Ensure the txs all end up in the pending queue by having the txs start
from nonce 0.

## Testing
Ran the benchmark and observed the full suite passing.

## Related Issues
Closes #1384.