feat(sequencer): have app recost transactions on fee or asset change (#1411)
This PR builds on top of the mempool balance aware logic to recost the
mempool when a `FeeAssetChange` or `FeeChange` action are seen in a
finalized block.

There needs to be a follow up PR to ensure that the `FeeChange` and
`FeeAssetChange` actions are only ran at the end of the block to ensure
that transactions are invalidated mid-block due to fee changes.

The mempool needs to recost transactions if the fees are changed, as the
original placement of the transaction in the `parked` or `pending`
structure may be out of date.

- Added logic to flag when `FeeAssetChange` or `FeeChange` actions are
included in a block.
- The mempool will now re-cost transactions when this happens.
- Changed the mempool to use the app's state directly instead of using
getters

Unit tests for the mempool and the app logic.

This doesn't have breaking changes, but the snapshot files were
re-generated because I changed the public addresses for `CAROL_ADDRESS`,
`BOB_ADDRESS`, and `JUDY_ADDRESS` in the `test_utils.rs` file. I needed
access to their private keys to sign transactions for testing promotion
flows on the app level.

Added `MEMPOOL_RECOSTED` counter to provide metrics on how often fees
are being changed.

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>