feat(sequencer): report deposit events (#1447)
## Summary
Added creation of ABCI deposit events and record them to the state to be
included in `ExecTxResult`.

## Background
Deposit events were previously only stored app-side in `SequencerBlock`.
This will allow for outside observation of the deposit events.

## Changes
- Added helper function for the creation of ABCI deposit events.
- Added recording of these events whenever deposits are created.

## Testing
Added test to ensure deposit events are correctly recorded and returned
by `state.apply()`

## Related Issues
closes #1440

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>
Co-authored-by: Fraser Hutchison <190532+Fraser999@users.noreply.github.com>
Co-authored-by: jesse snyder <jessetsnyder@gmail.com>