feat(sequencer): make mempool balance aware (#1408)
## Summary
Added logic to the mempool to make its pending vs parked logic balance
aware. Transactions will only be added to the pending queue if the
signing account has enough balance to cover the cost of all of the
transactions in the pending plus the new transaction. Transactions that
exceed the account's available balance are put into parked.

This PR also contains the logic needed to recost the transactions in the
mempool when a `FeeAssetChange` or `FeeChange` actions are ran.


![image](https://github.com/user-attachments/assets/a78851f3-81bd-4a42-b0af-46c4ad224f32)


## Background
Previously the mempool did not have sufficient logic to deal with
transaction costs. `CheckTx` would kick out any transaction that the
account did not have enough balance to cover. This didn't account for
the edge cases of:
1. In-flight transactions from other accounts funding the deficient
account.
2. Sequential transactions that together exceed an account's balance but
independently are fine. This causes unnecessary transaction failures in
`prepare_proposal()`'s block construction.

## Changes
- Added a `cost` field to the `TimemarkedTransaction` struct. 
- Made the `PendingTransactionsForAccount` type's `add()` functionality
reject transactions that would exceed an account's balance.
- Added cost-aware demotion logic to the `PendingTransactionsForAccount`
type and promotion logic to the `ParkedTransactionsForAccount` type
which is invoked in the mempool's `run_maintenance()` function.
- Factored out some functionality of internal functions to be used
elsewhere.
- Has the app flag when a `FeeAssetChange` or `FeeChange` actions is
executed and proceeds to recost the mempool in the `run_maintenance()`
function.

## Testing
Wrote unit tests and ran locally. 

## Metrics
Added timing metrics to CheckTx to track how long the transaction
costing logic takes:
- `check_tx_duration_seconds_convert_address`
- `check_tx_duration_seconds_fetch_balances`
- `check_tx_duration_seconds_fetch_tx_cost`
Added metrics to count how many times the mempool is recosted:
- `mempool_recosted`

## Related Issues
closes #1336

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>