chore(sequencer): migrate from `anyhow::Result` to `eyre::Result` (#1387)
## Summary
Migrate all instances of `anyhow::Result` to `eyre::Result`.

## Background
Sequencer was using `anyhow::Result`, which provides an unhelpful
`Display` impl and contrasts our error handling in the rest of the
codebase. This change is to flush out our error handling in the
sequencer, except for those parts which necessitate using `anyhow`.

## Changes
- Add eyre to sequencer's `cargo.toml`.
- Migrate all instances of `anyhow::Result` to `eyre::Result`, except
for those that touch cnidarium directly.
- Create `anyhow_to_eyre()` and `eyre_to_anyhow()` helper functions for
moving between the two without breaking the source chain.

## Testing
Added unit tests to ensure `eyre` and `anyhow` source chains are
maintained when converting between one another.

## Related Issues
closes #1386