ci(cli): astria-cli image for smoke tests (#1504)
## Summary
Adds a image build for the cli tool, and updates the justfiles to
utilize it.

## Background
There are two workflows which this fixes: 1) Users running the repo
locally and running smoke tests could encounter issues by not having the
cli installed or not having it updated. Now we just need to have docker.
Additionally, we have been needing to build the `just-cli` for each of
our smoke test runs. Instead all can now rely on the docker image.

## Changes
- Starts creating a new CLI image
- Updates the smoke tests to utilize CLI image in place of installing
the cli
- Update docker build scripts to make it easier to use a local image
locally

## Testing
CI/CD on this PR

## Related Issues

closes #1475