chore(sequencer)!: put blocks and deposits to non-verified storage (ENG-812) (#1525)
## Summary
This changes where sequencer blocks and the `TracePrefixed` native asset
are stored, and also persists `Deposit`s to non-verifiable storage.

## Background
Sequencer blocks are currently written to verified storage, which is
unnecessary as all of their constituent data is already available there.

We want to avoid the cost of storing full blocks in verified storage,
but still have the ability to provide a third party with a full block if
requested. To further assist with this, deposits will also be written to
non-verified storage.

It also seemed appropriate to store the `TracePrefixed` native asset in
verified storage, since this is something which all validators should
reach consensus over.

## Changes
- `SequencerBlock`s are now stored in non-verifiable storage.
- `Deposit`s are now stored in non-verifiable storage.
- The native asset is now stored in verifiable storage.

## Testing
Existing unit tests have been updated as required to assert these
changes.

## Breaking Changelist
- The set of data being written to verified storage has changed, which
is a breaking change in terms of on-disk data and generation of state
root hashes. Otherwise no APIs have changed.

## Related Issues
Closes #1493.