refactor(cli): merge argument parsing and command execution (#1568)
## Summary
Refactors the CLI to give it a cleaner structure.

## Background
The CLI was split into a `cli` module that was concerned with parsing
arguments, and a `commands` which consumed the argument data structures.

The patch clears this up by implementing inherent `<Args>::run` methods
on all leaf argument types, so that the implementation follows the same
structure for all its effects.

Note that this match does restructure the public facing command line
(except for a few lines of documentation). It merely restructures the
implementation.

## Changes
- Refactors each type deriving `clap::Args` to have an inherent method
`run`.
- Flattens the crate structure to not be unnecessarily nested.

## Testing
These will be tested in follow PRs. Smoke tests using a subset of the
astria CLI commands should still work.

## Related Issues
Closes https://github.com/astriaorg/astria/issues/1546