fix(sequencer): fix app hash in horcrux sentries (#1646)
## Summary
Changes the comparison when deciding if we need to re-execute a proposal
to use the validator address and the timestamp as identifiers, to enable
horcrux sentry nodes to not have app hash clashes.

## Background
When running horcrux all sentry nodes act as the same validator and all
nodes create a new proposal, but the signer only signs one of them. This
creates issues since all sentry nodes except the signed node will not
reexecute the proposal since it was created by their validator address
even though the proposals are different. Timestamps are used to add
additional fingerprinting to this without adding high computation cost.

## Changes
- Utilize timestamp as part of identifying whether to clear and execute
proposal in prepare proposal

## Testing
A validator reported this issue when running testnet, and similar
changes were tested and validated.