chore(proto)!: call transactions `Transaction`, contents `TransactionBody` (#1650)
## Summary
Renames `SignedTransaction` to just `Transaction`, `UnsignedTransaction`
to `TransactionBody`.

## Background
The differentiation between `SignedTransaction` and
`UnsignedTransaction` was confusing: an unsigned transaction does not
exist outside of sequencer, whereas the distinction into signed and
unsigned insinuates that both are valid and existential.

This patch renames `SignedTranscation` to just `Transaction`, and
`UnsignedTransaction` to `TransactionBody` to make it clear that it is
part of a `Transaction`.

## Changes
- Rename protobuf and domain type `SignedTransaction` to `Transaction`,
`UnsignedTransaction` to `TransactionBody`
- Rename `astria.protocol.transactions.v1alpha` to
`astria.protocol.transaction.v1alpha1` (plural to singular form)

## Testing
All changes compile, all tests run. Snapshot tests that involve the name
of the changed protobuf message are updated (as is the case for
`Transaction::id`, which constructs an identifier using the encoded
`Any` form of the inner transaction body). Tests for deposit events are
updated (because they use a sequence action's full name, which includes
the now changed package name).

## Breaking Changelist
- These changes are protocol breaking because a sequencer prior to this
patch will not be able to read a new transaction. And a sequencer after
this patch will not be able to read an old signed transaction (because
of the contained `Any` type).

## Related Issues
Closes https://github.com/astriaorg/astria/issues/1648