feat(sequencer): make ABCI response for account balances deterministic (#1574)
## Summary
Ensures that a response to an ABCI request for account balances is
deterministic.

## Background
Previously was not sorted, and the `AssetBalance` stream is
non-deterministic, which isn't great for an external API.

## Changes
- Created deterministic sorting function for sorting the `AssetBalance`s
firstly in descending order by balance, then in ascending order by name
if the balances are equal.

## Testing
Added unit test to ensure the sorting function works as expected.

## Related Issues
closes #1451