feat(bridge-withdrawer): add `use_compat_address` configuration value (#1671)
## Summary
add configuration option for `use_compat_address`, which can then be set
by the operator of the bridge withdrawer.

## Background
required for withdrawals from a rollup to noble, in case the withdrawal
on noble times out or fails.

## Changes
- add configuration option for `use_compat_address`

## Testing
existing unit tests, would need to have a withdrawal timeout test for
noble to fully test this

## Related Issues

closes #1424