ci: add test for rollup refund (#1728)
## Summary
Adds a test for case in which there is an IBC timeout after rollup
withdrawal to ensure that the funds are returned to the rollup. This is
done by using the `astria-cli` withdrawal functionality, after
collecting withdrawals we manually override the revision height to one
which is after the current sequencer chain consensus height but below
the destination chains current height. This triggers an automatic
instant timeout.

## Background
We previously had no e2e testing to ensure withdrawals were refunded
appropriately.

## Testing
Run against binaries which were known to have issues, as well as update
that should fix.