fix(core): replace todo by implementation (#1736)
## Summary
Actually implements the `Protobuf::to_raw` trait method on the
`TransactionBody` domain type.

## Background
`<TransactionBody as Protobuf>::to_raw` was not implemented, and instead
contained a `todo!()` macro. If called, this would panic at runtime
which is a problem.

At the time of writing this was not a problem in the codbase because the
method was not called. But it's very bad style to have that as part of a
public API.

## Changes
- Replace the inherent methods `TransactionBody::to_raw` and
`TransactionBody::into_raw` by the `Protobuf for TransactionBody` impl,
fixing the bad `todo!()` in its `Protobuf::to_raw` trait method.