fix(bridge-withdrawer, cli): use leading channel (#1768)
## Summary
Update the bridge contract parsing to utilize the correct portion of the
ibc channel for the source channel.

## Background
We were using the wrong channel which created errors with assets which
had more than one port/channel in ibc prefixing.

## Changes
- Removed `TracePrefixed::last_channel` (it is confusing and practivelly
never of interest to a consumer of the API).
- Added `TracePrefixed::leading_channel` to read the left-most channel
of an ICS20 asset.
- Updated all services to use `TracePrefixed::leading_channel`.

## Testing
CI testing + tested against prod environment to ensure correct output
with CLI

## Changelogs
Changelogs updated.

## Breaking Changelist
- Removes an API endpoint which was no longer used in astria-core.

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>