Apply new storage conventions to Connect oracle (#1774)
## Summary
This mainly brings the new oracle code in line with storage conventions
formed after the initial cut of the PR.

## Background
We want a consistent approach to storage across all components. As well
as using a single serialization format (Borsh), this will allow us to
iterate on the storage design in the near future.

## Changes
- Added `storage` submodules to the two sequencer modules
`connect::market_map` and `connect::oracle`. These provide the keys and
values which are used to store the relevant data.
- Added snapshot tests and other unit tests for the keys.
- Added put/get tests for the extension traits, which revealed bugs in
both the streaming getters of the oracle.
- Removed unneeded serde trait derives from several core types.
- Renamed `connect::marketmap` in the sequencer to `connect::market_map`

## Testing
Added several unit tests.

## Changelogs
We should update changelogs once the feature is ready to be merged to
`main`. Not done in this PR.

## Breaking Changelist
- This breaks the state snapshot tests due to changing how the oracle
data is serialized in storage.

---------

Co-authored-by: elizabeth <elizabethjbinks@gmail.com>
Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>