feat(sequencer, core): put extended commit info and proof into `SequencerBlock` (#1817)
## Summary
put extended commit info and merkle proof of inclusion into
`SequencerBlock`.

## Background
required for propagating oracle data inside the extended commit info to
rollups.

## Changes
- put extended commit info and merkle proof of inclusion into
`SequencerBlock` as optional types. if there is no extended commit info
for that block (ie. vote extensions are disabled) then they remain
unset. this is for backwards-compatibility
- also put put extended commit info and proof into celestia
`SubmittedMetadata` (so it is written to DA)

## Testing
unit tests

## Breaking Changelist
- `SequencerBlock` proto is updated with new fields; however, if these
fields are unset, the native `SequencerBlock` sets these to `None`, so
this branch should be able to sync with networks without vote extensions
enabled
- same for the `SubmittedMetadata` proto

## Related Issues

closes  #1764