run rollup streams off a single channel
also makes pushing into the duplex stream a method
on on the stream (rather than on handle to it).

moves conversion of optimistic to base block to
outside the channel so that it doesn't get
swallowed (the error event would be happening in
a long-running span, if under any span at all).