replace the joinset of live auctions by a single auction task
the running auctions were cancelled/aborted every
time a new optimistic block was received. This
means that no matter if the auction itself was not
finished because a block replacement occured or
because it could not be submitted to the mempool:
they would always end up being cancelled.

This means a single auction is all that's
necessary because no more than one auction future
at a time may be live.

This was implemented using a new RunningAction
type, that itself tracks its auction ID, the
spawned task and a channel to it, and a parent
block hash after having received an executed
block.

These changes mean that the CurrentBlock type
could be removed.