separate the trigger to start the auction timer from the bids
the trigger to start the auction timer comes from
the sequencer block commitment. the trigger to
start bids comes from the rollup executing the
optimistic block in time.

this means that the timer must start running
independently of whether or not the rollup returned
with the the executed block in time.