chore: update `url` dependency (#1869)
## Summary
Updated the version of  `url`.

## Background
Running `cargo audit` yields a warning of a vulnerability in `idna` (see
[RUSTSEC-2024-0421](https://rustsec.org/advisories/RUSTSEC-2024-0421)).
The suggested fix is to upgrade `idna` to >= v1.0.0.

`idna` is only a dependency in our tree of `url`, and we cannot upgrade
`idna` to >= v1.0.0 without upgrading `url`. However, upgrading `url`
from v2.5.2 to v2.5.4 causes the `idna` dependency to upgrade to v1.0.3,
hence fixing the issue.

## Changes
- Ran `cargo update -p url` to increase the version of `idna` to v1.0.3.

## Testing
Ran normal test suite.

## Changelogs
Changelogs updated.