chore(ci): action to autoclose stale issues and prs (#1862)
## Summary
Adds a regularly scheduled action to mark PRs and issues as stale, and
then autoclose with a message if no action taken.

## Background
Sometimes issues & pull requests stay around without action or cleanup.
By tagging stale items and checking them we can make sure the list of
PRs and issues we have is actually actionable. Closing does not impact
ability to find and use in future, but makes sure that pertinent issues
are easy to find.