chore(conductor): don't return empty tuple (#1885)
## Summary
Removed empty tuple member in return type.

## Background
We shouldn't be returning a tuple with an empty member.

## Changes
- Removed empty member from tuple return.

## Testing
Passing all tests.

## Changelogs
No updates required.

## Breaking Changes
Noting that this requires overriding the Conductor code freeze, but due
to the very small nature of this change I believe this will be okay
since it should not have any bearing on auditing of the service.

## Related Issues
closes #1879