chore(sequencer): remove misplaced logs
Sequencer checks if the storage path exists and
then reports on actions that it thinks what the
call `cnidarium::Storage::load` will do.

However, on the one hand cnidarium itself already
reports that information. On the other hand,
`Sequencer::run_until_stopped` is the wrong place
to emit these events because it does not have
authority over which actions cnidarium will take.
Should the construct ever change, then the events
will be immediately out of whack, which will lead
to confusion.