feat(chart): add auctioneer chart and just commands (#1902)
Adds a Helm chart for deploying the Auctioneer.

[PR #1556](https://github.com/astriaorg/astria/pull/1556) adds the
auctioneer binary. This binary has clients to both rollup and sequencer
nodes.

- Add the `charts/auctioneer` directory for the auctioneer binary
- Add commands to `charts/deploy.just` for deploying/deleting the
auctioneer chart in our k8s cluster
- Add `dev/auctioneer/values.yaml` for dev values for the auctioneer.
- Add the auctioneer's dev wallet to the genesis allocations in
`dev/validators/all.yaml` so that the auctioneer has a balance while
testing against the dev cluster.
- Expose the bundle grpc service from the rollup chart so that the
auctioneer can talk to it from within the cluster

Link any issues that are related, prefer full github links.

closes https://github.com/astriaorg/astria/issues/1766

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>

feat(charts): update evm-rollup charts for auctioneer (#1901)

Update the evm-rollup genesis charts to include
`astriaAuctioneerAddresses` which is a map of block height to auctioneer
addresses. We also add a new flag to geth called `auctioneer` which
enables auctioneer functionality in the flame node.

The auctioneer binary submits a signed bundle to the sequencer which the
flame side of the auctioneer unbundles and verifies. To verify the
bundle signature, the flame side requires the bech32 address of the
sequencer key which signed bundle. We pass the bech32 address of the
sequencer key to the flame side by adding it as a field in the genesis
file of the flame EVM rollup.

We also add a new flag in geth `auctioneer` which enables the auctioneer
funcationality in a geth node.

- Add a field `astriaAuctioneerAddresses` to the
`charts/evm-rollup/files/genesis/geth-genesis.json` which is
parameterized by `auctioneerAddresses`
- Add a field in the `charts/evm-rollup/values.yaml` and
`dev/rollup/dev.yaml` called `auctioneerAddresses`

By running the charts locally.

closes https://github.com/astriaorg/astria/issues/1829