feat(charts): update evm-rollup charts for auctioneer (#1901)
## Summary
Update the evm-rollup genesis charts to include
`astriaAuctioneerAddresses` which is a map of block height to auctioneer
addresses. We also add a new flag to geth called `auctioneer` which
enables auctioneer functionality in the flame node.

## Background
The auctioneer binary submits a signed bundle to the sequencer which the
flame side of the auctioneer unbundles and verifies. To verify the
bundle signature, the flame side requires the bech32 address of the
sequencer key which signed bundle. We pass the bech32 address of the
sequencer key to the flame side by adding it as a field in the genesis
file of the flame EVM rollup.

We also add a new flag in geth `auctioneer` which enables the auctioneer
funcationality in a geth node.

## Changes
- Add a field `astriaAuctioneerAddresses` to the
`charts/evm-rollup/files/genesis/geth-genesis.json` which is
parameterized by `auctioneerAddresses`
- Add a field in the `charts/evm-rollup/values.yaml` and
`dev/rollup/dev.yaml` called `auctioneerAddresses`

## Testing
By running the charts locally.

## Related Issues
closes https://github.com/astriaorg/astria/issues/1829