fix(sequencer)!: use bridge address to determine asset in bridge unlock cost estimation instead of signer (#1905)
## Summary
Changed bridge unlock cost calculation to use bridge address instead of
signer address.

## Background
Previously, the `get_total_transaction_cost` attempted to retrieve the
bridge account asset by the tx signer instead of by the bridge address.
This works fine if the sender is a bridge account, but if it is the
authorized bridge withdrawer, it will fail.

## Changes
- Change `BridgeUnlock` cost calculation to retrieve asset based on the
action's bridge address instead of the transaction signer.

## Testing
Added unit test to ensure function works correctly when submitting a
`BridgeUnlock` from the bridge withdrawer address.

Synced to mainnet from genesis to ensure this would not be a network
breaking change.

## Changelogs
Changelog updates.

## Breaking Changelist
- This change is breaking since previous `BridgeUnlock` submitted from
the withdrawer address would fail. This has been synced to mainnet
successfully from genesis as of January 13, 2025 at 2:32pm CST.

## Related Issues
closes #1904