refactor(sequencer): split run method to eliminate long-lived spans (#1898)
## Summary
Split `run_until_stopped()` into two methods and called them both in
`spawn()`.

## Background
There are two reasons for making this change, firstly being to eliminate
the long-lived span in `run_until_stopped()`. Secondly, this is for
parity with the rest of the codebase, which separates the creation
and/or setup of the components from the running process.

## Changes
- Moved initialization logic into `initialize()`, which is instrumented.
- Called both `initialize()` and `run_until_stopped()` in new method
`spawn()`.
- Moved signal handler initialization into `spawn()` and poll for signal
receipt during initialization.
- Added doc comment for the public `spawn()` method.

## Testing
Passing all current tests, no additional testing needed.

## Changelogs
Changelogs updated

## Breaking Changelist
- Breaks public API, as the public-facing run function is now `spawn()`
instead of `run_until_stopped()`

## Related Issues
closes #1895
closes #1893