fix(conductor): remove panic source on shutdown (#1919)
## Summary
Removes a panic source in the conductor frontend by moving around the
restart logic.

## Background
`Conductor::run_until_stopped` contains a loop to allow restarting its
backend business logic. This loop delegates to a
`Conductor::restart_or_shutdown` method to execute the restart of the
backend, and then immediately reenters the loop no matter a restart
happened or not.

While refactoring the inner conductor business logic (which lead to
slightly different return values in turn affecting the blackbox test
assertions) it was found that behavior can lead to the conductor backend
panicking because it's being polled again even after running to
completion.


## Changes
- Rename `Conductor::shutdown_or_restart` to
`Conductor::restart_or_shutdown` to match the enum it's branching on.
- Execute the restart inside the frontend loop to not reentering it,
avoiding to poll the already completed backend task.

## Testing 
This can't really be tested. But it does fix the panic in the refactor
PR.

## Changelogs
Changelogs updated.