feat(sequencer): add market map actions (#1882)
## Summary
Added select market map actions
[here](https://github.com/skip-mev/connect/blob/main/proto/connect/marketmap/v2/tx.proto)
to the sequencer.

## Background
These actions were needed for support for keeping a `MarketMap` in state
and making changes to it as part of the oracle feature. They follow
Skip's logic and apply it to our state machine.

#### `ChangeMarkets`
Takes a list of markets and either creates, updates, or removes them
depending on its variant. Must be signed by an address included in the
market map params' `market_authorities`.
- **Create:** Creates the markets in the market map. If no market map is
found, one will be created. If any of the markets to create already
exist, this action will err.
- **Update:** Updates the markets in the market map, matching based on
`Ticker.currency_pair`). If no market map is found, or any market is
missing a counterpart in the map, this action will err.
- **Remove:** Removes the markets from the market map. If a market is
not found in the map, it will be ignored.

#### `UpdateParams`
Updates the market map `Params`, which contains the market authority
addresses as well as an admin address. This will execute whether there
are params in the state already or not. Must be signed by the sequencer
network authority sudo address.

## Changes
- Added `ChangeMarkets` and `UpdateParams` to `PriceFeed` action in
proto, core, and sequencer.

## Testing
- Added fairly exhaustive unit tests to each action's implementation
module.
- Added app execution test for each action for more high-level testing.
- Added each action to all action snapshot.

## Breaking Changelist
- Added new actions which break app hash due to adding new information
to state.
- Added `ALICE_ADDRESS` to market authorities in testing `genesis`,
breaking two more snapshots.