feat(telemetry): emit span events for span construction, deletion (#1947)
## Summary
Instructs the tracing formatting subscriber to emit span events for
event construction and deletion.

## Background
Developers frequently add an event at the beginning of a function to get
information on whether a function is invoked. In tracing telemetry this
is strictly speaking superfluous because the existence of a span already
implies this information. If Astriia's services were primarily observed
via an opentelemetry platform, then this issue wouldn't arise in the
first place. But since events/logs are still consumed as plain text, we
should mimick this behavior by emitting events when a particular unit of
processing is started or finished.

Note that this will lead to noisier logs, but that is a justifiable
tradeoff because this information is frequently needed when debugging
services.

## Changes
- Change Astria's human readable trace/event output to constructs events
when spans are constructed or deleted (which usually means that an
instrumented function or method is invoked or finished).

## Testing
These changes can be observed when running a service locally.

## Changelogs
Changelog updated.