chore(sequencer): add missing storage key tests (#1989)
## Summary
Added missing `BridgeTransfer` fee storage key tests

## Background
These tests were a minor oversight in the implementation of
`BridgeTransfer`, but an easy fix to get in. As a followup, we should
get together some documentation for adding new actions such that there's
a quantitative list of items like this which are easy to miss.

## Changes
- Added storage key snapshot and fee component prefix test for
`BridgeTransfer` fees

## Testing
Passing tests

## Changelogs
No updates required

## Breaking Changelist
- Not breaking, just noting that this added a snapshot

## Related Issues
closes #1988