feat(conductor): include sequencer block hash (#1999)
## Summary
Adds a `sequencer_block_hash` to Execution API `Block` and
`ExecuteBlock` shapes such that an execution can include it in their
block space.

This is an addition but is backwards compatible. Old rollups can run the
new binary, rollups that don't use the sequencer hash can utilize the
old binary. Rollup forks which rely on the sequencer hash must utilize
the new binary.

## Background
We are upgrading Flame to the `Cancun` upgrade which includes
`BeaconRoot` in the block hash. This is a natural time to add the
sequencer block, as it is a similar concept.

## Changes
- Add new hash field to execution v1 protos and `astria-core`
- Added the field to in development execution v2 protos
- Update conductor to include the new data in `ExecuteBlock` requests
- Added information to the execute API spec.

## Testing
CI/CD testing has been updated to run the smoke tests with an updated
version of geth that utilizes the sequencer and includes a beacon root
contract in genesis.

I have done manual testing of the following:
- old geth w/ new conductor unchanged
- new geth w/ cancun upgrade disabled

## Changelogs
Changelogs updated

## Breaking Changelist
This is not a breaking change as it is a protobuf addition.