fix(sequencer): mempool pending nonce calcuations (#2012)
## Summary
Updates when the mempool is rebalanced in order to solve for race
condition in which call falls through to state before it has been
written, by rebalancing after commit and updates the mempool methods to
return the pending account nonce instead of the highest nonce in the
pending mempool.

## Background
Nonce is defined as the number of transactions which have been executed
on the account. The pending nonce should then be the number of
transactions executed + the number of transactions sitting in pending.
Our API was not doing this correctly. Pending nonce on the mempool was
being defined as the transaction with the highest nonce as opposed to
what the account nonce will be after all pending transactions.

Additionally, our mempool was being rebalanced before state was
committed, this can result in a race condition where you request the
pending nonce after rebalancing has occured but before committed state
has occured resulting in fetching not the pending nonce but the current
nonce.

## Changes
- Move mempool maintenance to after `commit` from finalized block
- Mempool defines pending nonce as pending account nonce instead of
highest nonce in queue

## Testing
CI/CD tests, which have been updated.

## Changelogs
Changelogs updated.

## Related Issues
closes #2011