fix(sequencer): ensure proposal execution uses latest info (#2024)
## Summary
Replaces the proposal fingerprint and and executed block hash with an
`ExecutionState` statemachine object which tracks that we move between
acceptable execution pipelines.

## Background
We had a system which used a proposal fingerprint and an executed block
hash. One was for determining if we should execute during process
proposal and the other was for execution in finalize block. These two
were linked but with seperate logic. The fingerprint wasn't adequate on
the proposal.

## Changes
- merged `executed_app_hash` and `executed_proposal_fingerprint` into
the `ExecutionState` object.
- Added more fields into how we fingerprint a proposal.

## Testing
CI/CD tests updated, added more detailed test cases. Unit tests for the
ExecutionState

## Changelogs
Changelogs updated.

## Related Issues
closes https://github.com/astriaorg/astria/issues/1835
closes https://github.com/astriaorg/internal-planning/issues/13