fix(sequencer): increase mempool removal cache size (#1969)
## Summary
This changes the hard-coded value of the mempool removal cache size to
be greater than the default CometBFT mempool size.

## Background
The removal cache is the means by which CometBFT becomes aware that a
failed tx should be removed from its own mempool (via a CheckTx call).
If the removal cache reaches the limit of entries, new ones will replace
the oldest ones before CometBFT is able to call CheckTx for them.

Note that the removal cache entries are fairly small - a couple of
hashes and a small "removal reason" object each, so increasing this
limit significantly should have no meaningful effect on the sequencer's
memory usage or performance.

## Changes
- Increase `REMOVAL_CACHE_SIZE` from 4096 to 50,000.

## Testing
No new tests added.

## Changelogs
Changelogs updated.

## Related Issues
Closes #1968.