chore(core, sequencer): simplify price feed action (#2113)
## Summary
The new price feed action has been flattened into two new top-level
actions, and authorization for these has been simplified to exclusively
require the sudo address.

## Background
The action was structured in a similar way to corresponding objects in
Skip's Connect project. We want to flatten the action into the parent
`Action` to keep it more in line with other existing actions.

We also want to simplify authorization for executing these actions.
Again we followed the design in Skip Connect, but this is not required
in Astria, at least for now.

## Changes
- Removed `PriceFeed` action and price feed `Params` types.
- Made `CurrencyPairsChange` into a top-level action.
- Renamed `ChangeMarkets` to `MarketsChange` and made this into a
top-level action.
- Renamed `MarketsChange` variants to `creation`, `removal` and
`update`,
- Changed auth checks for these to require the tx signer be the sudo
address.

## Testing
Existing tests were sufficient - these have just been refactored to
match the new types.

## Changelogs
No updates required - these are covered by the existing entries around
addition of price feed data.

## Breaking Changelist
- Despite changing some API types, these types have not yet been
published, so this is effectively not a breaking change.

## Related Issues
Closes #2112.