fix(sequencer): remove failed promotable instead of inserted tx (#2135)
## Summary
Fixes an error in mempool promotion logic after insertion.

## Background
Previously, any transactions which failed to be added to the `Pending`
queue during promotion post-insertion would remove the just-inserted
transaction ID from `contained_txs`. This is an error, but it hasn't
caused any problems since this code path is meant to never happen.
Additionally, any transactions which failed adding to pending should be
added to the CometBFT removal cache.

## Changes
- In `Mempool::insert`, post-insertion, in failure case of adding
promoted transaction to `pending`: remove `tx_id` from `contained_txs`
instead of the transaction which was successfully added.
- Add failing transaction to CometBFT removal cache.
- Added `RemovalReason::InternalError`.

## Testing
Added unit test to ensure this functionality works as expected. The test
scenario is quite contrived, but without direct manipulation of the
`pending` and `parked` caches, we have no other way of creating this
edge case.

## Changelogs
Changelog updated.

## Related Issues
closes <!-- list any issues closed here --> #2130