feat(core, sequencer): provide checked actions and transaction (#2142)
## Summary
This introduces `CheckedAction`, `CheckedTransaction` and `Checked...`
wrappers for each action type.

## Background
We want to have stricter checks performed when `CheckTx` is called on a
new tx in order to avoid accepting txs which subsequently fail
execution.

Checks on txs and their included actions are now split into two
categories: mutable and immutable. On construction of a checked object,
both types of checks are run. Subsequently, only the mutable checks are
re-run, and this is currently only done at the start of executing the
action/transaction.

Nonce and balance checks are excluded from these checks as the Mempool
only puts forward for execution txs which should have valid nonces and
sufficient account balances.

In order to ensure such checks have been performed, the sequencer now
deals almost exclusively in checked flavours of actions and txs.

As a result of this, many tests which previously used txs had to be
refactored to ensure that constructing the equivalent checked txs would
succeed.

## Changes
- Introduced new `Checked...` types for actions and `Transaction`.
Replaced usage throughout sequencer codebase of the unchecked flavours
of these.
- Changed `ExpandedBlockData::user_submitted_transactions` to UN-parsed
transactions so that all checking can be done when constructing a
`CheckedTransaction`.
- Unified several test utilities into a single, top-level `test_utils`
module. As well as pre-existing helper functions, there is now a
`Fixture` for intializing a new chain along with several
builders/initializers to aid in constructing objects to be used in
tests. These are provided with doc comments which should explain how to
use them.
- Disabled benchmarks for now. These are already not functioning on
MacOS (albeit there is an open PR #1842 to address this). I would prefer
to handle updating the benchmarks as part of that PR rather than adding
to this one.

## Testing
- Existing tests were generally refactored to retain their original
intent.
- Individual checked actions were provided with extensive unit test
coverage. Each one generally tests the construction AND execution of the
specific action. To generate failures during execution, the unit tests
construct the action to be tested while the global state makes
construction possible (e.g signer is sudo), and then modify the global
state to make execution fail (e.g. change the sudo address). All such
state changes are done by executing other checked actions rather than
directly manipulating global state, to ensure that such state changes
remain feasible in the future.
- These new action unit tests made the tests in
`tests_execute_transaction.rs` largely redundant, hence that file has
been deleted. Coverage was also provided for `IbcRelay` where none
existed before, but the setup difficulty for testing `Ics20Withdrawal`
is such that only a subset of the code paths are tested (as is the case
currently).
- `CheckedTransaction` also has extensive unit test coverage for
construction and execution, and covers any deleted tests from
`tests_execute_transaction.rs` not already catered for by the action
unit tests.

## Changelogs
Changelogs updated.

## Metrics
- Added metrics:
  - `ASTRIA_SEQUENCER_CHECK_TX_FAILED_ACTION_CHECKS`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CHECK_ACTIONS`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_RECHECK`
- Renamed metric `ASTRIA_SEQUENCER_CHECK_TX_REMOVED_TOO_LARGE` to
  `ASTRIA_SEQUENCER_CHECK_TX_FAILED_TX_TOO_LARGE`
- Deleted metrics:
  - `ASTRIA_SEQUENCER_CHECK_TX_REMOVED_FAILED_STATELESS`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_PARSE_TX`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CHECK_STATELESS`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CHECK_TRACKED`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CHECK_CHAIN_ID`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CHECK_REMOVED`
  - `ASTRIA_SEQUENCER_CHECK_TX_DURATION_SECONDS_CONVERT_ADDRESS`

## Breaking Changelist
- There should be no consensus, API or state-breaking changes other than
the metrics listed above. All checks being done on actions and
transactions remain the same, they are simply performed at a different
stage of the flow now.

## Related Issues
Closes #1958.
Closes #1620.