chore(ci): add `mempool-apis` and `upgrades` to proto lint (#2166)
## Summary
Adds the `mempool-apis` and `upgrades` proto packages to the `proto` job
in the `lint` workflow.

## Background
The `mempool-apis` and `upgrades` packages were added recently, but
neither have been added to the lint since they were not yet added in
BUF. Now that they are, this adds these packages to the breaking changes
lint.

## Changes
- Adds `mempoolapis` and `upgrades` to the `lint/proto` job.

## Testing
CI passing.

## Changelogs
No updates required.

## Related Issues
Closes #2145