chore(sequencer): update handling of addition and removal actions (#2171)
## Summary
This changes how actions involving addition or removal of data are
handled.

## Background
In some cases currently, we silently ignore additions where the data
already exists, or removals where the data does not exist. We want to
always treat these as error conditions.

## Changes
- Changed `CurrencyPairsChange::Addition` action to fail if any currency
pair to be added is already stored.
- Changed `CurrencyPairsChange::Removal` action to fail if any currency
pair to be removed is not currently stored.
- Changed `FeeAssetChange::Addition` action to fail if the fee asset to
be added is already stored.
- Changed `FeeAssetChange::Removal` action to fail if the fee asset to
be removed is not currently stored.
- Changed `IbcRelayerChange::Addition` action to fail if the address to
be added is already stored.
- Changed `IbcRelayerChange::Removal` action to fail if the address to
be removed is not currently stored.
- Changed `MarketsChange::Removal` action to fail if any market to be
removed is not currently stored.

## Testing
Added/updated unit tests for all of the changed behaviour.

## Changelogs
Changelogs updated.

## Related Issues
Closes #2158.