feat(sequencer): add transaction fee gRPC query (#2160)
## Summary
Adds transaction fee gRPC query to the Mempool API.

## Background
We have an ABCI query for fees; now that we are shifting towards gRPC
usage it would be good to have a gRPC endpoint for obtaining a
transaction's fees as well. This change is meant to address this.

## Changes
- Added `GetTransactionFeeRequest` and `GetTransactionFeeResponse` to
`mempoolapis`.
- Added `GetTransactionFees` to `TransactionService`.
- Added implementation of `get_transaction_fees` in the mempool server,
utilizing the same logic as in `fees::query`.

## Testing
Added one failure case unit test and one success case test.

## Changelogs
Changelog updated.

## Related Issues
closes #2152