feat(sequencer)!: add recently included tx result cache to mempool (#2163)
## Summary
Adds a cache of recently included transactions' execution results to the
mempool.

## Background
#2133 added support for getting a transaction's status from the mempool.
When CometBFT calls `CheckTx`, however, transactions are removed from
the removal cache. So, transactions which have been included in a block
have an extremely limited amount of time to be observed as "included"
before they disappear from the mempool completely. This is bad UX, and
these changes aim to address that.

## Changes
- Adds `RecentExecutionResults` cache to mempool with a timeout of 1
minute and a maximum of `ASTRIA_SEQUENCER_EXECUTION_RESULTS_CACHE_SIZE`
results included.

## Testing
- Added unit tests for `RecentExecutionResults`
- Added unit tests in mempool to ensure `transaction_status` correctly
grabs the status from `recent_execution_results`.

## Changelogs
Changelog updated.

## Breaking Changes
- Added `ASTRIA_SEQUENCER_EXECUTION_RESULTS_CACHE_SIZE` to env vars.

## Metrics
- Added gauge `results_in_recently_executed_cache`.

## Related Issues
closes #2147