feat(charts)!: flame major upgrade (#2138)
## Summary
Major chart upgrades for Flame and its deployment.

## Background
The concept of Forks, extradata override changes, an update of the
deposit tx id, the new deployment of Flame node using an app of app
pattern, and new images are the reason these chart updates are required.

## Changes
- Genesis file updates (for all networks)
- Chart updates for Prometheus rules and image updates 

## Testing
The new images and apps were deployed on devnet and testnet, and new
nodes were synced. App state hashes were then verified to have not
changed between the new deployment and the old deployment.

## Breaking Changelist
- The addition of forks
- The deposit tx id change 

## Related Issues
https://github.com/astriaorg/argocd-apps/pull/328